refactor: descomponer build_dev_db.py en paquete src/builders/
El God module de build_dev_db.py (2867 lineas) se descompone en 9 modulos
cohesivos bajo src/builders/ y main() se separa en subfunciones orquestadas,
dejando build_dev_db.py en ~670 lineas de solo orquestacion.

Modulos nuevos: _shared, io_utils, formats, metadata, reports, artifacts,
datasets, catalog, landing. main() se divide en _load_inputs,
_compute_validations, _write_data_artifacts, build_dataset_metadata y
_generate_reports.

DATASET_CATALOG_CONFIG y las funciones publicas se re-exportan desde
build_dev_db para no romper scripts/verify_pipeline.py ni los tests.

Verificacion: 410 tests pasan, lint limpio, y el pipeline end-to-end produce
artefactos byte-identicos a la linea base salvo timestamps. Se corrigio un
return perdido en attach_publishable_package_to_manifest (sin cobertura de
tests en el camino feliz de main). check_validation_registration.py se ajusta
para tolerar la nueva asignacion validations = _compute_validations(...).

Docs actualizados (scorecard, NEXT_STEPS, ficha 01, AGENTS.md, SOURCE_OF_TRUTH.md).
