+++ src/build_dev_db.py
@@ -1,15 +1,9 @@
-import hashlib
 import json
 import os
-import re
-import sqlite3
 import sys
-import zipfile
-from datetime import datetime, timezone
+from datetime import timezone
 
-import duckdb
 import polars as pl
-import tomllib
 
 UTC = timezone.utc
 
@@ -17,9 +11,88 @@ ROOT_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 if ROOT_DIR not in sys.path:
     sys.path.insert(0, ROOT_DIR)
 
+# Rutas, constantes y configuración de catálogo compartidas (re-exportadas para
+# mantener compatibilidad con scripts externos y tests que las importan desde aquí;
+# DATASET_CATALOG_CONFIG se consume vía build_dev_db en verify_pipeline.py y tests).
+from src.builders._shared import (  # noqa: E402, F401
+    CENSO_HOGARES_METADATA_PATH,
+    CENSO_METADATA_PATH,
+    COMUNAS_METADATA_PATH,
+    DATASET_CATALOG_CONFIG,
+    ELECTORAL_METADATA_PATH,
+    EMPRESAS_METADATA_PATH,
+    FINANZAS_METADATA_PATH,
+    INDICADORES_METADATA_PATH,
+    NORMALIZED_DIR,
+    RESULTADOS_EDUCACIONALES_METADATA_PATH,
+    SALUD_METADATA_PATH,
+    SIEDU_METADATA_PATH,
+    STAGING_DIR,
+)
+from src.builders.artifacts import (  # noqa: E402
+    attach_publishable_package_to_manifest,
+    write_artifact_manifest,
+    write_hub_bundle_json,
+    write_publishable_bundle_sha256,
+    write_publishable_bundle_zip,
+)
+from src.builders.catalog import (  # noqa: E402
+    write_dataset_catalog,
+    write_pipeline_metadata,
+)
+from src.builders.datasets import (  # noqa: E402
+    build_perfil_territorial_comunal,
+    derive_geography_layers,
+)
+from src.builders.formats import (  # noqa: E402
+    build_duckdb,
+    build_excel,
+    build_flat_files,
+    build_sqlite,
+)
+from src.builders.io_utils import (  # noqa: E402
+    write_json_atomic,
+    write_parquet_atomic,
+)
+
+# build_coverage, build_degradation y build_drift se re-exportan para los tests
+# (no se usan directamente aquí, solo a través de enrich_dataset_metadata).
+from src.builders.metadata import (  # noqa: E402, F401
+    build_coverage,
+    build_dataset_metadata,
+    build_degradation,
+    build_drift,
+    build_freshness,
+    build_freshness_warnings,
+    build_indicator_delivery,
+    enrich_dataset_metadata,
+    ensure_directories,
+    load_metadata,
+    load_schema_contract,
+)
+from src.builders.reports import (  # noqa: E402
+    build_dataset_changelog,
+    build_dataset_quality,
+    build_dataset_status,
+    build_drift_report,
+    build_hub_status,
+    build_overview,
+    build_provenance_report,
+    build_redistribution_report,
+    build_source_readiness,
+    write_dataset_changelog_json,
+    write_dataset_quality_json,
+    write_dataset_status_json,
+    write_drift_report_json,
+    write_hub_health_json,
+    write_hub_status_json,
+    write_overview_json,
+    write_provenance_report_json,
+    write_redistribution_report_json,
+    write_source_readiness_json,
+)
 from src.pipeline_status_utils import (
     build_hub_health,
-    compute_freshness,
     write_dataset_catalog_markdown_file,
     write_dataset_quality_markdown_file,
     write_drift_report_markdown_file,
@@ -47,2145 +120,9 @@ from src.validation import (
     validate_resultados_educacionales,
 )
 
-# Configuración de rutas
-DATA_DIR = os.path.abspath(os.path.join(os.path.dirname(__file__), "../data"))
-STAGING_DIR = os.path.join(DATA_DIR, "staging")
-NORMALIZED_DIR = os.path.join(DATA_DIR, "normalized")
-COMUNAS_METADATA_PATH = os.path.join(STAGING_DIR, "comunas.metadata.json")
-INDICADORES_METADATA_PATH = os.path.join(STAGING_DIR, "indicadores.metadata.json")
-CENSO_METADATA_PATH = os.path.join(STAGING_DIR, "censo_comunal.metadata.json")
-SALUD_METADATA_PATH = os.path.join(STAGING_DIR, "establecimientos_salud.metadata.json")
-CENSO_HOGARES_METADATA_PATH = os.path.join(STAGING_DIR, "censo_hogares_viviendas.metadata.json")
-ELECTORAL_METADATA_PATH = os.path.join(STAGING_DIR, "distritos_electorales.metadata.json")
-FINANZAS_METADATA_PATH = os.path.join(STAGING_DIR, "finanzas_municipales.metadata.json")
-RESULTADOS_EDUCACIONALES_METADATA_PATH = os.path.join(
-    STAGING_DIR, "resultados_educacionales.metadata.json"
-)
-SIEDU_METADATA_PATH = os.path.join(STAGING_DIR, "indicadores_urbanos_siedu.metadata.json")
-EMPRESAS_METADATA_PATH = os.path.join(STAGING_DIR, "empresas.metadata.json")
-EXCEL_MAX_ROWS = 1_048_576  # Límite físico de Excel por hoja
-PUBLISHABLE_ARTIFACT_SUFFIXES = (".json", ".md", ".parquet")
-PUBLISHABLE_BUNDLE_ZIP_NAME = "chile-hub-publishable-bundle.zip"
-PUBLISHABLE_BUNDLE_SHA256_NAME = "chile-hub-publishable-bundle.zip.sha256"
 
-
-def _load_catalog_config() -> dict:
-    path = os.path.join(ROOT_DIR, "data", "dataset_catalog_config.json")
-    if not os.path.isfile(path):
-        raise FileNotFoundError(
-            f"Archivo de configuración de catálogo no encontrado: {path}\n"
-            "Este archivo es la fuente de verdad para los metadatos de datasets.\n"
-            "Asegúrate de que existe en el repositorio (debe estar tracked en git).\n"
-            "Si acabas de clonar, verifica que el archivo no esté en .gitignore."
-        )
-    with open(path, "r", encoding="utf-8") as f:
-        return json.load(f)
-
-
-DATASET_CATALOG_CONFIG = _load_catalog_config()
-
-
-def build_freshness(refreshed_at_utc, max_age_hours):
-    return compute_freshness(refreshed_at_utc, max_age_hours)
-
-
-def build_freshness_warnings(dataset_name, freshness):
-    status = freshness.get("status")
-    age_hours = freshness.get("age_hours")
-    max_age_hours = freshness.get("max_age_hours")
-
-    if status == "stale":
-        return [
-            (
-                f"{dataset_name} freshness is stale: "
-                f"{age_hours}h since refresh, policy max is {max_age_hours}h"
-            )
-        ]
-    if status == "unknown":
-        return [f"{dataset_name} freshness is unknown: missing or invalid refreshed_at_utc"]
-    return []
-
-
-def build_indicator_delivery(metadata):
-    indicator_codes = metadata.get("indicator_codes", [])
-    delivery = metadata.get("indicator_delivery")
-    if isinstance(delivery, dict) and delivery:
-        return delivery
-
-    delivery = {code: "live" for code in indicator_codes}
-    for pair in metadata.get("raw_recoveries", []):
-        delivery[pair.split("/", 1)[0]] = "raw_recovery"
-    for pair in metadata.get("preserved_existing_pairs", []):
-        delivery[pair.split("/", 1)[0]] = "preserved_existing"
-    for code in metadata.get("published_backfills", []):
-        delivery[code] = "published_backfill"
-    return delivery
-
-
-def build_degradation(dataset_name, dataset_metadata, validation):
-    source_mode = dataset_metadata.get("source_mode")
-    record_count = dataset_metadata.get("record_count")
-    warnings = validation.get("warnings", [])
-
-    if source_mode == "fallback":
-        if dataset_name == "comunas":
-            return {
-                "status": "degraded",
-                "impact": (
-                    f"Cobertura territorial parcial: {record_count} comunas disponibles "
-                    "desde fallback embebido."
-                ),
-                "recommended_action": "Reintentar extractores o restaurar la fuente territorial primaria.",
-            }
-        if dataset_name in {"regiones", "provincias"}:
-            return {
-                "status": "degraded",
-                "impact": (
-                    f"Capa derivada desde comunas en fallback; cardinalidad reducida a {record_count} filas."
-                ),
-                "recommended_action": "Recuperar comunas live para restaurar cobertura derivada completa.",
-            }
-        if dataset_name == "indicadores":
-            return {
-                "status": "degraded",
-                "impact": "Valores provenientes de fallback local; no representan el último snapshot live.",
-                "recommended_action": "Reintentar la API pública antes de redistribuir o usar en reporting.",
-            }
-
-    if warnings:
-        return {
-            "status": "warning",
-            "impact": "; ".join(warnings),
-            "recommended_action": "Revisar warnings operativos del dataset antes de consumirlo en producción.",
-        }
-
-    return {
-        "status": "none",
-        "impact": "Sin degradación operativa detectada en este build.",
-        "recommended_action": "Ninguna.",
-    }
-
-
-def build_coverage(dataset_name, dataset_metadata):
-    declared_coverage = dataset_metadata.get("coverage", {})
-    if declared_coverage.get("status") == "partial_expected":
-        return {
-            "status": "partial",
-            "expected_record_count": None,
-            "actual_record_count": dataset_metadata.get("record_count"),
-            "coverage_ratio": declared_coverage.get("coverage_ratio"),
-            "summary": declared_coverage.get(
-                "expected_scope",
-                "Cobertura parcial esperada y declarada por la fuente.",
-            ),
-        }
-
-    expected_record_count = DATASET_CATALOG_CONFIG.get(dataset_name, {}).get(
-        "expected_record_count"
-    )
-    actual_record_count = dataset_metadata.get("record_count")
-
-    if expected_record_count is None:
-        return {
-            "status": "not_applicable",
-            "expected_record_count": None,
-            "actual_record_count": actual_record_count,
-            "coverage_ratio": None,
-            "summary": "Sin baseline de cobertura por cardinalidad para esta capa.",
-        }
-
-    if actual_record_count is None or expected_record_count <= 0:
-        return {
-            "status": "unknown",
-            "expected_record_count": expected_record_count,
-            "actual_record_count": actual_record_count,
-            "coverage_ratio": None,
-            "summary": "No fue posible calcular cobertura contra el baseline esperado.",
-        }
-
-    coverage_ratio = round(actual_record_count / expected_record_count, 4)
-    if actual_record_count >= expected_record_count:
-        status = "full"
-        summary = (
-            f"Cobertura completa: {actual_record_count}/{expected_record_count} filas "
-            "respecto del baseline esperado."
-        )
-    else:
-        status = "partial"
-        summary = (
-            f"Cobertura parcial: {actual_record_count}/{expected_record_count} filas "
-            "respecto del baseline esperado."
-        )
-
-    return {
-        "status": status,
-        "expected_record_count": expected_record_count,
-        "actual_record_count": actual_record_count,
-        "coverage_ratio": coverage_ratio,
-        "summary": summary,
-    }
-
-
-def build_drift(dataset_metadata):
-    source_mode = dataset_metadata.get("source_mode")
-    coverage = dataset_metadata.get("coverage", {})
-    degradation = dataset_metadata.get("degradation", {})
-    coverage_status = coverage.get("status", "unknown")
-    degradation_status = degradation.get("status", "none")
-
-    drift_status = "healthy"
-    if (
-        source_mode == "fallback"
-        or coverage_status in {"partial", "unknown"}
-        or degradation_status in {"warning", "degraded"}
-    ):
-        drift_status = "drifted"
-
-    if drift_status == "healthy":
-        summary = "Sin drift operativo detectado en este build."
-        recommended_action = "Ninguna."
-    else:
-        summary = (
-            f"Drift detectado: mode={source_mode}, coverage={coverage_status}, "
-            f"degradation={degradation_status}."
-        )
-        recommended_action = degradation.get(
-            "recommended_action",
-            "Revisar fuente, cobertura y warnings antes de consumir esta capa.",
-        )
-
-    return {
-        "status": drift_status,
-        "summary": summary,
-        "recommended_action": recommended_action,
-    }
-
-
-def enrich_dataset_metadata(dataset_metadata, validations):
-    enriched = {}
-    for dataset_name, metadata in dataset_metadata.items():
-        validation = validations.get(dataset_name, {})
-        degradation = build_degradation(dataset_name, metadata, validation)
-        coverage = build_coverage(dataset_name, metadata)
-        # Incrustar campos del contrato de esquema para comparación de changelog
-        contract = load_schema_contract(dataset_name)
-        contract_fields = {}
-        if contract:
-            contract_fields = {
-                "contract_primary_key": contract.get("primary_key", []),
-                "contract_required_columns": contract.get("required_columns", []),
-                "contract_column_types": contract.get("column_types", {}),
-                "contract_nullable_columns": contract.get("nullable_columns", []),
-                "contract_exists": True,
-            }
-        enriched[dataset_name] = {
-            **metadata,
-            "degradation": degradation,
-            "coverage": coverage,
-            **contract_fields,
-        }
-        enriched[dataset_name]["drift"] = build_drift(enriched[dataset_name])
-    return enriched
-
-
-def load_schema_contract(dataset_name):
-    """Carga el contrato de esquema para un dataset.
-
-    Retorna un dict con los campos del contrato, o None si el archivo
-    no existe o no se puede interpretar.
-    """
-    contract_path = os.path.join(ROOT_DIR, "contracts", "datasets", f"{dataset_name}.schema.json")
-    if not os.path.exists(contract_path):
-        return None
-    try:
-        with open(contract_path, "r", encoding="utf-8") as f:
-            return json.load(f)
-    except (json.JSONDecodeError, OSError):
-        return None
-
-
-def ensure_directories():
-    os.makedirs(NORMALIZED_DIR, exist_ok=True)
-
-
-def validate_metadata_schema(path, content):
-    """
-    Valida que el archivo JSON de metadatos de staging no sea nulo, sea un JSON válido
-    y contenga todos los campos obligatorios del contrato de datos.
-    """
-    required_fields = {
-        "dataset",
-        "source_name",
-        "source_url",
-        "source_mode",
-        "refreshed_at_utc",
-        "record_count",
-        "fields",
-        "reuse_policy",
-    }
-    if not isinstance(content, dict):
-        raise SystemExit(
-            f"Error: Metadatos en {os.path.relpath(path, ROOT_DIR)} no es un objeto JSON (dict)."
-        )
-
-    missing_fields = required_fields - set(content.keys())
-    if missing_fields:
-        raise SystemExit(
-            f"Error: Metadatos en {path} no cumple con el esquema obligatorio.\n"
-            f"Campos faltantes: {', '.join(sorted(missing_fields))}"
-        )
-
-
-def load_metadata(path):
-    if not os.path.exists(path):
-        return None
-    try:
-        with open(path, encoding="utf-8") as f:
-            content = json.load(f)
-    except json.JSONDecodeError as e:
-        raise SystemExit(
-            f"Error: Archivo de metadatos {os.path.relpath(path, ROOT_DIR)} "
-            f"contiene un JSON malformado: {e}"
-        )
-
-    validate_metadata_schema(path, content)
-    return content
-
-
-def write_json_atomic(data, path, **kwargs):
-    tmp_path = path + ".tmp"
-    with open(tmp_path, "w", encoding="utf-8") as f:
-        json.dump(data, f, **kwargs)
-        f.write("\n")
-    os.replace(tmp_path, path)
-
-
-def sync_landing_metadata(public_site_url):
-    index_path = os.path.join(ROOT_DIR, "index.html")
-    app_path = os.path.join(ROOT_DIR, "app.js")
-    public_site_url = public_site_url.rstrip("/") + "/"
-    public_data_base = public_site_url + "data/normalized"
-
-    DISPLAY_NAMES = {
-        "regiones": "Regiones de Chile",
-        "provincias": "Provincias de Chile",
-        "comunas": "División Político-Administrativa (DPA) de Chile",
-        "comunas_enriquecidas": "Comunas Enriquecidas de Chile",
-        "indicadores": "Indicadores Económicos de Chile",
-        "censo_comunal": "Censo Comunal de Chile 2024",
-        "censo_hogares_viviendas": "Censo Hogares y Viviendas de Chile 2024",
-        "establecimientos_salud": "Establecimientos de Salud de Chile",
-        "distritos_electorales": "Distritos Electorales de Chile",
-        "establecimientos_educacionales": "Establecimientos Educacionales de Chile",
-        "finanzas_municipales": "Finanzas Municipales de Chile",
-        "resultados_educacionales": "Resultados Educacionales de Chile",
-        "indicadores_urbanos_siedu": "Indicadores Urbanos SIEDU de Chile",
-        "perfil_territorial_comunal": "Perfil Territorial Comunal de Chile",
-        "empresas": "Registro de Empresas y Sociedades de Chile (RES)",
-    }
-
-    CREATORS = {
-        "regiones": "Biblioteca del Congreso Nacional de Chile (BCN)",
-        "provincias": "Biblioteca del Congreso Nacional de Chile (BCN)",
-        "comunas": "Biblioteca del Congreso Nacional de Chile (BCN)",
-        "comunas_enriquecidas": "Biblioteca del Congreso Nacional de Chile (BCN) e Instituto Nacional de Estadísticas (INE)",
-        "indicadores": "Banco Central de Chile e Instituto Nacional de Estadísticas (INE)",
-        "censo_comunal": "Instituto Nacional de Estadísticas (INE)",
-        "censo_hogares_viviendas": "Instituto Nacional de Estadísticas (INE)",
-        "establecimientos_salud": "Ministerio de Salud de Chile (MINSAL)",
-        "distritos_electorales": "Biblioteca del Congreso Nacional de Chile (BCN) y Servicio Electoral de Chile (SERVEL)",
-        "establecimientos_educacionales": "Ministerio de Educación de Chile (MINEDUC)",
-        "finanzas_municipales": "Sistema Nacional de Información Municipal (SINIM) y SUBDERE",
-        "resultados_educacionales": "Ministerio de Educación de Chile (MINEDUC)",
-        "indicadores_urbanos_siedu": "Instituto Nacional de Estadísticas (INE) y SIEDU",
-        "perfil_territorial_comunal": "chile-hub (Capa derivada a partir de datasets validados de chile-hub)",
-        "empresas": "Registro de Empresas y Sociedades (RES) - Ministerio de Economía de Chile",
-    }
-
-    try:
-        if os.path.exists(index_path):
-            with open(index_path, "r", encoding="utf-8") as f:
-                content = f.read()
-
-            datasets_json_ld = []
-            for key, config in DATASET_CATALOG_CONFIG.items():
-                name = DISPLAY_NAMES.get(key, key)
-                desc = config.get("description", "")
-                reuse_policy = config.get("reuse_policy", {})
-                license_url = reuse_policy.get(
-                    "license_url", "https://creativecommons.org/licenses/by/4.0/"
-                )
-                creator_name = CREATORS.get(key, "chile-hub")
-
-                datasets_json_ld.append(
-                    {
-                        "@type": "Dataset",
-                        "name": name,
-                        "description": desc,
-                        "url": f"{public_site_url}#dataset-{key}",
-                        "license": license_url,
-                        "creator": {"@type": "Organization", "name": creator_name},
-                    }
-                )
-
-            catalog_json_ld = {
-                "@context": "https://schema.org",
-                "@type": "DataCatalog",
-                "name": "chile-hub",
-                "url": public_site_url,
-                "description": "Capa de datos oficial, curada y reproducible sobre Chile. Datasets limpios en Parquet, JSON, DuckDB y Excel.",
-                "publisher": {"@type": "Organization", "name": "chile-hub"},
-                "dataset": datasets_json_ld,
-            }
-
-            json_ld_string = json.dumps(catalog_json_ld, indent=2, ensure_ascii=False)
-            indented_json_ld = "\n".join(
-                "    " + line if line.strip() else "" for line in json_ld_string.splitlines()
-            )
-
-            pattern_marker = (
-                r"<!-- START_DATA_CATALOG_JSON_LD -->.*?<!-- END_DATA_CATALOG_JSON_LD -->"
-            )
-            new_script = f'<!-- START_DATA_CATALOG_JSON_LD -->\n    <script type="application/ld+json">\n{indented_json_ld}\n    </script>\n    <!-- END_DATA_CATALOG_JSON_LD -->'
-
-            if "<!-- START_DATA_CATALOG_JSON_LD -->" in content:
-                new_content = re.sub(pattern_marker, new_script, content, flags=re.DOTALL)
-            else:
-                pattern_catalog = r'<script type="application/ld\+json">\s*\{\s*"@context":\s*"https://schema\.org",\s*"@type":\s*"DataCatalog".*?</script>'
-                new_content = re.sub(pattern_catalog, new_script, content, flags=re.DOTALL)
-
-            replacements = [
-                (
-                    r"https://cortega26\.github\.io/chile-hub/",
-                    public_site_url,
-                ),
-            ]
-            for pattern, replacement in replacements:
-                new_content = re.sub(pattern, replacement, new_content)
-
-            if new_content != content:
-                with open(index_path, "w", encoding="utf-8") as f:
-                    f.write(new_content)
-                print(
-                    f"Sincronización Landing: index.html actualizado a {public_site_url} con JSON-LD sincronizado."
-                )
-        if os.path.exists(app_path):
-            with open(app_path, "r", encoding="utf-8") as f:
-                content = f.read()
-            new_content = re.sub(
-                r'const PUBLIC_DATA_BASE = "[^"]+";',
-                f'const PUBLIC_DATA_BASE = "{public_data_base}";',
-                content,
-            )
-            if new_content != content:
-                with open(app_path, "w", encoding="utf-8") as f:
-                    f.write(new_content)
-                print(f"Sincronización Landing: app.js actualizado a {public_data_base}")
-    except Exception as e:
-        print(f"Advertencia: No se pudo actualizar la landing: {e}")
-
-
-def write_pipeline_metadata(dataset_metadata, validations):
-    version = "unknown"
-    public_site_url = "https://tooltician.com/chile-hub/"
-    try:
-        pyproject_path = os.path.join(ROOT_DIR, "pyproject.toml")
-        with open(pyproject_path, "rb") as f:
-            pyproject_data = tomllib.load(f)
-        version = pyproject_data.get("project", {}).get("version", "unknown")
-        public_site_url = (
-            pyproject_data.get("tool", {})
-            .get("chile_hub", {})
-            .get("public_site_url", public_site_url)
-        )
-    except Exception as e:
-        print(f"Advertencia: No se pudo obtener la versión de pyproject.toml: {e}")
-
-    if version != "unknown":
-        sync_landing_metadata(public_site_url)
-
-    pipeline_metadata = {
-        "version": version,
-        "public_site_url": public_site_url,
-        "generated_at_utc": datetime.now(UTC).isoformat(),
-        "datasets": dataset_metadata,
-        "validations": validations,
-    }
-    output_path = os.path.join(NORMALIZED_DIR, "pipeline_metadata.json")
-    write_json_atomic(pipeline_metadata, output_path, ensure_ascii=False, indent=2)
-    return output_path, pipeline_metadata
-
-
-def write_dataset_catalog(pipeline_metadata):
-    datasets = []
-    for dataset_name, dataset_metadata in pipeline_metadata["datasets"].items():
-        validation = pipeline_metadata["validations"].get(dataset_name, {})
-        config = DATASET_CATALOG_CONFIG.get(dataset_name, {})
-        freshness_policy = config.get("freshness_policy", {})
-        datasets.append(
-            {
-                "dataset": dataset_name,
-                "description": config.get("description", ""),
-                "alias_for": config.get("alias_for"),
-                "source_name": dataset_metadata.get("source_name"),
-                "source_url": dataset_metadata.get("source_url"),
-                "source_mode": dataset_metadata.get("source_mode"),
-                "source_detail": dataset_metadata.get("source_detail"),
-                "refreshed_at_utc": dataset_metadata.get("refreshed_at_utc"),
-                "record_count": dataset_metadata.get("record_count"),
-                "fields": dataset_metadata.get("fields", []),
-                "indicator_codes": dataset_metadata.get("indicator_codes", []),
-                "indicator_delivery": dataset_metadata.get("indicator_delivery", {}),
-                "join_keys": config.get("join_keys", []),
-                "confidence_tier": config.get("confidence_tier"),
-                "reuse_policy": config.get("reuse_policy", {}),
-                "freshness": dataset_metadata.get("freshness", {}),
-                "freshness_policy": freshness_policy,
-                "coverage": dataset_metadata.get("coverage", {}),
-                "drift": dataset_metadata.get("drift", {}),
-                "usage_examples": config.get("usage_examples", {}),
-                "outputs": config.get("outputs", {}),
-                "documentation": config.get("documentation"),
-                "validation_status": validation.get("status"),
-                "warnings": validation.get("warnings", []),
-                "degradation": dataset_metadata.get("degradation", {}),
-                "notes": dataset_metadata.get("notes", []),
-            }
-        )
-
-    catalog = {
-        "generated_at_utc": pipeline_metadata["generated_at_utc"],
-        "dataset_count": len(datasets),
-        "datasets": datasets,
-    }
-    output_path = os.path.join(NORMALIZED_DIR, "dataset_catalog.json")
-    write_json_atomic(catalog, output_path, ensure_ascii=False, indent=2)
-    return output_path, catalog
-
-
-def compute_sha256(path):
-    hasher = hashlib.sha256()
-    with open(path, "rb") as f:
-        for chunk in iter(lambda: f.read(65536), b""):
-            hasher.update(chunk)
-    return hasher.hexdigest()
-
-
-def build_publishable_artifact_index():
-    artifact_index = {}
-    # Load source registry to filter by public_bundle_eligible
-    registry = load_source_registry()
-    registry_by_dataset = {entry["dataset"]: entry for entry in registry}
-
-    for dataset_name, config in DATASET_CATALOG_CONFIG.items():
-        # Los datasets con alias_for son punteros de compatibilidad que comparten
-        # los mismos archivos físicos que el dataset canónico. Saltarlos aquí evita
-        # que sobrescriban el mapeo de artefactos y deja que el dataset canónico
-        # sea el dueño de sus archivos.
-        if config.get("alias_for"):
-            continue
-
-        reg_entry = registry_by_dataset.get(dataset_name, {})
-        public_bundle_eligible = reg_entry.get("public_bundle_eligible", True)
-        publication_track = reg_entry.get("publication_track", "stable_publishable")
-
-        # Only include datasets eligible for the public bundle
-        if not isinstance(public_bundle_eligible, bool) or not public_bundle_eligible:
-            continue
-
-        outputs = config.get("outputs", {})
-        for output_type, path in outputs.items():
-            if isinstance(path, str) and path.startswith("data/normalized/"):
-                artifact_index[path] = {
-                    "dataset": dataset_name,
-                    "output_type": output_type,
-                    "publication_track": publication_track,
-                    "public_bundle_eligible": True,
-                }
-    shared_artifacts = {
-        "data/normalized/pipeline_metadata.json": {
-            "shared_type": "pipeline_metadata",
-            "format": "json",
-        },
-        "data/normalized/pipeline_status.md": {
-            "shared_type": "pipeline_status",
-            "format": "markdown",
-        },
-        "data/normalized/hub_health.json": {
-            "shared_type": "hub_health",
-            "format": "json",
-        },
-        "data/normalized/hub_health.md": {
-            "shared_type": "hub_health",
-            "format": "markdown",
-        },
-        "data/normalized/hub_status.json": {
-            "shared_type": "hub_status",
-            "format": "json",
-        },
-        "data/normalized/dataset_status.json": {
-            "shared_type": "dataset_status",
-            "format": "json",
-        },
-        "data/normalized/dataset_changelog.json": {
-            "shared_type": "dataset_changelog",
-            "format": "json",
-        },
-        "data/normalized/hub_bundle.json": {
-            "shared_type": "hub_bundle",
-            "format": "json",
-        },
-        "data/normalized/redistribution_report.json": {
-            "shared_type": "redistribution_report",
-            "format": "json",
-        },
-        "data/normalized/redistribution_report.md": {
-            "shared_type": "redistribution_report",
-            "format": "markdown",
-        },
-        "data/normalized/provenance_report.json": {
-            "shared_type": "provenance_report",
-            "format": "json",
-        },
-        "data/normalized/provenance_report.md": {
-            "shared_type": "provenance_report",
-            "format": "markdown",
-        },
-        "data/normalized/drift_report.json": {
-            "shared_type": "drift_report",
-            "format": "json",
-        },
-        "data/normalized/drift_report.md": {
-            "shared_type": "drift_report",
-            "format": "markdown",
-        },
-        "data/normalized/overview.json": {
-            "shared_type": "overview",
-            "format": "json",
-        },
-        "data/normalized/overview.md": {
-            "shared_type": "overview",
-            "format": "markdown",
-        },
-        "data/normalized/dataset_catalog.json": {
-            "shared_type": "dataset_catalog",
-            "format": "json",
-        },
-        "data/normalized/dataset_catalog.md": {
-            "shared_type": "dataset_catalog",
-            "format": "markdown",
-        },
-        "data/normalized/artifact_manifest.json": {
-            "shared_type": "artifact_manifest",
-            "format": "json",
-        },
-        "data/normalized/source_readiness.json": {
-            "shared_type": "source_readiness",
-            "format": "json",
-        },
-        "data/normalized/source_readiness.md": {
-            "shared_type": "source_readiness",
-            "format": "markdown",
-        },
-        "data/normalized/dataset_quality.json": {
-            "shared_type": "dataset_quality",
-            "format": "json",
-        },
-        "data/normalized/dataset_quality.md": {
-            "shared_type": "dataset_quality",
-            "format": "markdown",
-        },
-    }
-    artifact_index.update(shared_artifacts)
-    return artifact_index
-
-
-def write_artifact_manifest():
-    artifact_index = build_publishable_artifact_index()
-    artifacts = []
-    for filename in sorted(os.listdir(NORMALIZED_DIR)):
-        if not filename.endswith(PUBLISHABLE_ARTIFACT_SUFFIXES):
-            continue
-        path = os.path.join(NORMALIZED_DIR, filename)
-        if not os.path.isfile(path):
-            continue
-        relative_path = f"data/normalized/{filename}"
-        artifact_metadata = artifact_index.get(relative_path)
-        # Only include artifacts that have metadata in the index.
-        # Candidate dataset artifacts are excluded from the index and will be skipped.
-        if artifact_metadata is None:
-            continue
-        artifacts.append(
-            {
-                "path": relative_path,
-                "dataset": artifact_metadata.get("dataset"),
-                "output_type": artifact_metadata.get("output_type"),
-                "shared_type": artifact_metadata.get("shared_type"),
-                "format": artifact_metadata.get("format"),
-                "size_bytes": os.path.getsize(path),
-                "sha256": compute_sha256(path),
-            }
-        )
-
-    manifest = {
-        "generated_at_utc": datetime.now(UTC).isoformat(),
-        "artifact_count": len(artifacts),
-        "artifacts": artifacts,
-        "packages": [],
-    }
-    output_path = os.path.join(NORMALIZED_DIR, "artifact_manifest.json")
-    write_json_atomic(manifest, output_path, ensure_ascii=False, indent=2)
-    return output_path, manifest
-
-
-def write_hub_health_json(health):
-    output_path = os.path.join(NORMALIZED_DIR, "hub_health.json")
-    write_json_atomic(health, output_path, ensure_ascii=False, indent=2)
-    return output_path
-
-
-def build_hub_status(hub_health):
-    return {
-        "generated_at_utc": hub_health.get("generated_at_utc"),
-        "overall_status": hub_health.get("overall_status"),
-        "dataset_count": hub_health.get("dataset_count"),
-        "live_count": hub_health.get("live_count"),
-        "fallback_count": hub_health.get("fallback_count"),
-        "stale_count": hub_health.get("stale_count"),
-        "drifted_count": hub_health.get("drifted_count"),
-        "degraded_count": hub_health.get("degraded_count"),
-        "warning_count": hub_health.get("warning_count"),
-        "top_issue": hub_health.get("top_issue"),
-        "top_issue_summary": hub_health.get("top_issue_summary"),
-    }
-
-
-def write_hub_status_json(hub_status):
-    output_path = os.path.join(NORMALIZED_DIR, "hub_status.json")
-    write_json_atomic(hub_status, output_path, ensure_ascii=False, indent=2)
-    return output_path
-
-
-def build_dataset_status(pipeline_metadata):
-    datasets = []
-    for dataset_name, dataset in sorted(pipeline_metadata.get("datasets", {}).items()):
-        validation = pipeline_metadata.get("validations", {}).get(dataset_name, {})
-        config = DATASET_CATALOG_CONFIG.get(dataset_name, {})
-        datasets.append(
-            {
-                "dataset": dataset_name,
-                "validation_status": validation.get("status"),
-                "source_mode": dataset.get("source_mode"),
-                "freshness": dataset.get("freshness", {}),
-                "record_count": dataset.get("record_count"),
-                "expected_record_count": config.get("expected_record_count"),
-                "coverage_status": dataset.get("coverage", {}).get("status"),
-                "coverage_ratio": dataset.get("coverage", {}).get("coverage_ratio"),
-                "redistribution_status": config.get("reuse_policy", {}).get("status"),
-                "redistribution_ok": config.get("reuse_policy", {}).get("redistribution_ok"),
-                "refreshed_at_utc": dataset.get("refreshed_at_utc"),
-                "warnings": validation.get("warnings", []),
-                "recommended_action": dataset.get("drift", {}).get(
-                    "recommended_action", "Ninguna."
-                ),
-            }
-        )
-    return {
-        "generated_at_utc": pipeline_metadata.get("generated_at_utc"),
-        "dataset_count": len(datasets),
-        "datasets": datasets,
-    }
-
-
-def write_dataset_status_json(dataset_status):
-    output_path = os.path.join(NORMALIZED_DIR, "dataset_status.json")
-    write_json_atomic(dataset_status, output_path, ensure_ascii=False, indent=2)
-    return output_path
-
-
-def _is_compatible_type_change(old_type, new_type):
-    """Determina si un cambio de tipo de columna es compatible hacia atrás.
-
-    El mismo tipo siempre es compatible.
-    Widening integer → float es compatible.
-    Cualquier otro cambio de tipo es incompatible.
-    """
-    if old_type == new_type:
-        return True
-    if old_type == "integer" and new_type == "float":
-        return True
-    return False
-
-
-def _compute_dataset_change_severity(current, previous, entry, is_new_dataset):
-    """Calcula los campos de severidad para un dataset comparando contratos.
-
-    Usa los campos contract_* incrustados en pipeline_metadata por
-    enrich_dataset_metadata(). Maneja el caso de migración donde previous
-    no tiene campos de contrato.
-    """
-    breaking_changes = []
-    new_columns = []
-    removed_columns = []
-    primary_key_changed = False
-    contract_changed = False
-
-    prev_has_contract = previous.get("contract_exists", False)
-    curr_has_contract = current.get("contract_exists", False)
-
-    if prev_has_contract and curr_has_contract:
-        prev_pk = previous.get("contract_primary_key", [])
-        curr_pk = current.get("contract_primary_key", [])
-        prev_required = set(previous.get("contract_required_columns", []))
-        curr_required = set(current.get("contract_required_columns", []))
-        prev_types = previous.get("contract_column_types", {})
-        curr_types = current.get("contract_column_types", {})
-        prev_nullable = set(previous.get("contract_nullable_columns", []))
-        curr_nullable = set(current.get("contract_nullable_columns", []))
-
-        # 1. Clave primaria cambiada → major
-        if prev_pk != curr_pk:
-            primary_key_changed = True
-            contract_changed = True
-            breaking_changes.append(f"Primary key changed from {prev_pk} to {curr_pk}")
-
-        # 2. Columnas requeridas eliminadas → major
-        removed_required = prev_required - curr_required
-        for col in sorted(removed_required):
-            breaking_changes.append(f"Required column removed: {col}")
-            contract_changed = True
-
-        # 3. Cambios de tipo incompatibles → major
-        for col, prev_type in prev_types.items():
-            curr_type = curr_types.get(col)
-            if curr_type and prev_type != curr_type:
-                if not _is_compatible_type_change(prev_type, curr_type):
-                    breaking_changes.append(
-                        f"Column type changed incompatibly: {col} from {prev_type} to {curr_type}"
-                    )
-                    contract_changed = True
-
-        # 4. Columnas agregadas o eliminadas del contrato
-        prev_contract_cols = set(prev_types.keys()) | prev_required | prev_nullable
-        curr_contract_cols = set(curr_types.keys()) | curr_required | curr_nullable
-        added_cols = curr_contract_cols - prev_contract_cols
-        removed_cols = prev_contract_cols - curr_contract_cols
-
-        for col in sorted(added_cols):
-            if col in curr_nullable:
-                new_columns.append(col)
-            else:
-                breaking_changes.append(f"New non-nullable column added: {col}")
-                contract_changed = True
-        if any(col in curr_nullable for col in added_cols):
-            contract_changed = True
-
-        for col in sorted(removed_cols):
-            removed_columns.append(col)
-            breaking_changes.append(f"Column removed from contract: {col}")
-            contract_changed = True
-
-    # Determinar severidad global
-    if breaking_changes:
-        change_severity = "major"
-    elif contract_changed:
-        change_severity = "minor"
-    elif is_new_dataset:
-        change_severity = "minor"
-    elif (
-        entry.get("added_fields")
-        or entry.get("removed_fields")
-        or entry.get("record_count_delta", 0) != 0
-    ):
-        change_severity = "patch"
-    else:
-        change_severity = "none"
-
-    return {
-        "change_severity": change_severity,
-        "breaking_changes": breaking_changes,
-        "new_columns": new_columns,
-        "removed_columns": removed_columns,
-        "primary_key_changed": primary_key_changed,
-        "contract_changed": contract_changed,
-    }
-
-
-def build_dataset_changelog(current_metadata, previous_metadata=None):
-    previous_datasets = (previous_metadata or {}).get("datasets", {})
-    current_datasets = current_metadata.get("datasets", {})
-    changes = []
-    for dataset_name in sorted(current_datasets):
-        current = current_datasets[dataset_name]
-        previous = previous_datasets.get(dataset_name, {})
-        current_fields = set(current.get("fields", []))
-        previous_fields = set(previous.get("fields", []))
-
-        entry = {
-            "dataset": dataset_name,
-            "previous_record_count": previous.get("record_count"),
-            "current_record_count": current.get("record_count"),
-            "record_count_delta": (
-                current.get("record_count") - previous.get("record_count")
-                if isinstance(current.get("record_count"), int)
-                and isinstance(previous.get("record_count"), int)
-                else None
-            ),
-            "added_fields": sorted(current_fields - previous_fields),
-            "removed_fields": sorted(previous_fields - current_fields),
-            "previous_source_mode": previous.get("source_mode"),
-            "current_source_mode": current.get("source_mode"),
-            "previous_freshness_status": previous.get("freshness", {}).get("status"),
-            "current_freshness_status": current.get("freshness", {}).get("status"),
-            "previous_validation_status": (previous_metadata or {})
-            .get("validations", {})
-            .get(dataset_name, {})
-            .get("status"),
-            "current_validation_status": current_metadata.get("validations", {})
-            .get(dataset_name, {})
-            .get("status"),
-        }
-
-        # Calcular severidad comparando contratos actual y anterior
-        is_new_dataset = dataset_name not in previous_datasets
-        severity_info = _compute_dataset_change_severity(current, previous, entry, is_new_dataset)
-        entry.update(severity_info)
-
-        changes.append(entry)
-
-    return {
-        "generated_at_utc": current_metadata.get("generated_at_utc"),
-        "previous_generated_at_utc": (previous_metadata or {}).get("generated_at_utc"),
-        "dataset_count": len(changes),
-        "datasets": changes,
-    }
-
-
-def write_dataset_changelog_json(dataset_changelog):
-    output_path = os.path.join(NORMALIZED_DIR, "dataset_changelog.json")
-    write_json_atomic(dataset_changelog, output_path, ensure_ascii=False, indent=2)
-    return output_path
-
-
-def build_redistribution_report(dataset_catalog):
-    datasets = []
-    for entry in dataset_catalog.get("datasets", []):
-        reuse_policy = entry.get("reuse_policy", {})
-        redistribution_ok = reuse_policy.get("redistribution_ok")
-        publishability_status = (
-            "ready"
-            if redistribution_ok is True
-            else "review_terms"
-            if redistribution_ok is False
-            else "unknown"
-        )
-        if publishability_status == "ready":
-            recommended_action = "Publicable con atribucion y referencia de fuente."
-        elif publishability_status == "review_terms":
-            recommended_action = "Revisar terminos vigentes antes de redistribuir fuera del repo."
-        else:
-            recommended_action = "Aclarar licencia o terminos antes de redistribuir."
-
-        datasets.append(
-            {
-                "dataset": entry.get("dataset"),
-                "publishability_status": publishability_status,
-                "license": reuse_policy.get("license"),
-                "license_url": reuse_policy.get("license_url"),
-                "reuse_status": reuse_policy.get("status"),
-                "attribution_required": reuse_policy.get("attribution_required"),
-                "redistribution_ok": redistribution_ok,
-                "summary": reuse_policy.get("summary"),
-                "recommended_action": recommended_action,
-                "documentation": entry.get("documentation"),
-            }
-        )
-
-    report = {
-        "generated_at_utc": dataset_catalog.get("generated_at_utc"),
-        "dataset_count": len(datasets),
-        "ready_count": sum(1 for entry in datasets if entry["publishability_status"] == "ready"),
-        "review_terms_count": sum(
-            1 for entry in datasets if entry["publishability_status"] == "review_terms"
-        ),
-        "unknown_count": sum(
-            1 for entry in datasets if entry["publishability_status"] == "unknown"
-        ),
-        "datasets": datasets,
-    }
-    return report
-
-
-def write_redistribution_report_json(report):
-    output_path = os.path.join(NORMALIZED_DIR, "redistribution_report.json")
-    write_json_atomic(report, output_path, ensure_ascii=False, indent=2)
-    return output_path
-
-
-def build_provenance_report(dataset_catalog):
-    datasets = []
-    for entry in dataset_catalog.get("datasets", []):
-        freshness = entry.get("freshness", {})
-        freshness_status = freshness.get("status", "unknown")
-        age_hours = freshness.get("age_hours")
-        max_age_hours = freshness.get("max_age_hours")
-        if age_hours is None or max_age_hours is None:
-            freshness_label = freshness_status
-        else:
-            freshness_label = f"{freshness_status} ({age_hours}h / {max_age_hours}h)"
-        warnings = entry.get("warnings", [])
-        notes = entry.get("notes", [])
-        diagnostic_summary = "Sin observaciones operativas."
-        if warnings:
-            diagnostic_summary = warnings[0]
-        elif notes:
-            diagnostic_summary = notes[0]
-
-        datasets.append(
-            {
-                "dataset": entry.get("dataset"),
-                "source_name": entry.get("source_name"),
-                "source_url": entry.get("source_url"),
-                "source_mode": entry.get("source_mode"),
-                "source_detail": entry.get("source_detail"),
-                "refreshed_at_utc": entry.get("refreshed_at_utc"),
-                "freshness_status": freshness_status,
-                "freshness_label": freshness_label,
-                "reuse_status": entry.get("reuse_policy", {}).get("status"),
-                "warning_count": len(warnings),
-                "diagnostic_summary": diagnostic_summary,
-                "documentation": entry.get("documentation"),
-            }
-        )
-
-    return {
-        "generated_at_utc": dataset_catalog.get("generated_at_utc"),
-        "dataset_count": len(datasets),
-        "live_count": sum(1 for entry in datasets if entry.get("source_mode") == "live"),
-        "fallback_count": sum(1 for entry in datasets if entry.get("source_mode") == "fallback"),
-        "datasets": datasets,
-    }
-
-
-def write_provenance_report_json(report):
-    output_path = os.path.join(NORMALIZED_DIR, "provenance_report.json")
-    write_json_atomic(report, output_path, ensure_ascii=False, indent=2)
-    return output_path
-
-
-def build_drift_report(dataset_catalog):
-    datasets = []
-    for entry in dataset_catalog.get("datasets", []):
-        source_mode = entry.get("source_mode")
-        coverage = entry.get("coverage", {})
-        degradation = entry.get("degradation", {})
-        drift = entry.get("drift", {})
-        coverage_status = coverage.get("status", "unknown")
-        degradation_status = degradation.get("status", "none")
-        drift_status = drift.get("status", "healthy")
-        warnings = entry.get("warnings", [])
-        diagnostic_summary = "Sin observaciones operativas."
-        if warnings:
-            diagnostic_summary = warnings[0]
-        elif degradation.get("impact"):
-            diagnostic_summary = degradation.get("impact")
-
-        datasets.append(
-            {
-                "dataset": entry.get("dataset"),
-                "drift_status": drift_status,
-                "source_mode": source_mode,
-                "coverage_status": coverage_status,
-                "coverage_ratio": coverage.get("coverage_ratio"),
-                "coverage_summary": coverage.get("summary"),
-                "degradation_status": degradation_status,
-                "degradation_impact": degradation.get("impact"),
-                "drift_summary": drift.get("summary"),
-                "warning_count": len(warnings),
-                "diagnostic_summary": diagnostic_summary,
-                "recommended_action": drift.get("recommended_action"),
-                "documentation": entry.get("documentation"),
-            }
-        )
-
-    return {
-        "generated_at_utc": dataset_catalog.get("generated_at_utc"),
-        "dataset_count": len(datasets),
-        "drifted_count": sum(1 for entry in datasets if entry["drift_status"] == "drifted"),
-        "healthy_count": sum(1 for entry in datasets if entry["drift_status"] == "healthy"),
-        "fallback_count": sum(1 for entry in datasets if entry["source_mode"] == "fallback"),
-        "partial_coverage_count": sum(
-            1 for entry in datasets if entry["coverage_status"] == "partial"
-        ),
-        "degraded_count": sum(1 for entry in datasets if entry["degradation_status"] == "degraded"),
-        "datasets": datasets,
-    }
-
-
-def write_drift_report_json(report):
-    output_path = os.path.join(NORMALIZED_DIR, "drift_report.json")
-    write_json_atomic(report, output_path, ensure_ascii=False, indent=2)
-    return output_path
-
-
-def build_overview(hub_health, hub_bundle, artifact_manifest):
-    primary_package = next(
-        (
-            package
-            for package in artifact_manifest.get("packages", [])
-            if package.get("package_type") == "zip"
-        ),
-        None,
-    )
-    return {
-        "generated_at_utc": hub_health.get("generated_at_utc"),
-        "overall_status": hub_health.get("overall_status"),
-        "dataset_count": hub_health.get("dataset_count"),
-        "live_count": hub_health.get("live_count"),
-        "fallback_count": hub_health.get("fallback_count"),
-        "stale_count": hub_health.get("stale_count"),
-        "drifted_count": hub_health.get("drifted_count"),
-        "degraded_count": hub_health.get("degraded_count"),
-        "partial_coverage_count": hub_health.get("partial_coverage_count"),
-        "warning_count": hub_health.get("warning_count"),
-        "top_issue": hub_health.get("top_issue"),
-        "top_issue_summary": hub_health.get("top_issue_summary"),
-        "shared_artifact_count": len(
-            [entry for entry in artifact_manifest.get("artifacts", []) if entry.get("shared_type")]
-        ),
-        "package_count": len(artifact_manifest.get("packages", [])),
-        "primary_package": (
-            {
-                "path": primary_package.get("path"),
-                "package_type": primary_package.get("package_type"),
-                "size_bytes": primary_package.get("size_bytes"),
-                "checksum_algorithm": primary_package.get("checksum_algorithm"),
-                "checksum_path": primary_package.get("checksum_path"),
-                "verification_command": primary_package.get("verification_command"),
-            }
-            if primary_package
-            else None
-        ),
-        "report_keys": sorted(hub_bundle.get("reports", {}).keys()),
-        "datasets": [
-            {
-                "dataset": entry.get("dataset"),
-                "source_mode": entry.get("source_mode"),
-                "validation_status": entry.get("validation_status"),
-                "freshness_status": entry.get("freshness_status"),
-                "coverage_status": entry.get("coverage_status"),
-                "drift_status": entry.get("drift_status"),
-            }
-            for entry in hub_health.get("datasets", [])
-        ],
-    }
-
-
-def write_overview_json(overview):
-    output_path = os.path.join(NORMALIZED_DIR, "overview.json")
-    write_json_atomic(overview, output_path, ensure_ascii=False, indent=2)
-    return output_path, overview
-
-
-def load_source_registry():
-    """Carga el registry de fuentes desde data/source_registry.json."""
-    registry_path = os.path.join(DATA_DIR, "source_registry.json")
-    with open(registry_path, "r", encoding="utf-8") as f:
-        return json.load(f)
-
-
-def build_source_readiness(pipeline_metadata):
-    """Construye el reporte de madurez de fuente combinando registry y metadatos."""
-    registry = load_source_registry()
-    datasets_metadata = pipeline_metadata.get("datasets", {})
-
-    entries = []
-    for src in registry:
-        dataset_name = src["dataset"]
-        ds_meta = datasets_metadata.get(dataset_name, {})
-
-        # Derivar URL del contrato de esquema
-        contract_path = f"contracts/datasets/{dataset_name}.schema.json"
-        contract_abs = os.path.join(ROOT_DIR, contract_path)
-        source_contract_url = (
-            f"github:chile-hub/chile-hub/{contract_path}" if os.path.exists(contract_abs) else None
-        )
-
-        # Determinar estancamiento
-        review_by = src.get("review_by")
-        stalled_after_days = src.get("stalled_after_days", 90)
-        stalled = False
-        if review_by:
-            try:
-                review_date = datetime.fromisoformat(review_by).replace(tzinfo=UTC)
-                stalled = datetime.now(UTC) > review_date
-            except (ValueError, TypeError):
-                pass
-
-        # Acción recomendada
-        recommended_action = src.get("next_action", "—")
-        if stalled:
-            recommended_action = f"⚠ ESTANCADO (revisión vencida {review_by}). {recommended_action}"
-
-        entries.append(
-            {
-                "dataset": dataset_name,
-                "maturity_status": src.get("maturity_status", "unknown"),
-                "source_id": src.get("source_id", ""),
-                "source_name": src.get("source_name", ""),
-                "official_url": src.get("official_url", ""),
-                "access_method": src.get("access_method", ""),
-                "license_status": src.get("license_status", ""),
-                "source_mode": ds_meta.get("source_mode", "not_built"),
-                "source_detail": ds_meta.get("source_detail", ""),
-                "live_ready": src.get("live_ready", False),
-                "fallback_allowed": src.get("fallback_allowed", False),
-                "publish_blocking": src.get("publish_blocking", False),
-                "live_extractor_status": src.get("live_extractor_status", "unknown"),
-                "fallback_policy": src.get("fallback_policy", "none"),
-                "source_contract_url": source_contract_url,
-                "stalled_after_days": stalled_after_days,
-                "review_by": review_by,
-                "stalled": stalled,
-                "owner": src.get("owner", "core"),
-                "next_action": src.get("next_action", "—"),
-                "recommended_action": recommended_action,
-            }
-        )
-
-    maturity_counts = {
-        s: sum(1 for e in entries if e["maturity_status"] == s)
-        for s in ("stable", "candidate", "experimental", "deprecated")
-    }
-
-    return {
-        "generated_at_utc": pipeline_metadata.get(
-            "generated_at_utc", datetime.now(UTC).isoformat()
-        ),
-        "dataset_count": len(entries),
-        **{f"{k}_count": v for k, v in maturity_counts.items()},
-        "live_ready_count": sum(1 for e in entries if e["live_ready"]),
-        "fallback_only_count": sum(
-            1 for e in entries if e["live_extractor_status"] == "fallback_only"
-        ),
-        "publish_blocking_count": sum(1 for e in entries if e["publish_blocking"]),
-        "datasets": entries,
-    }
-
-
-def write_source_readiness_json(readiness):
-    output_path = os.path.join(NORMALIZED_DIR, "source_readiness.json")
-    write_json_atomic(readiness, output_path, ensure_ascii=False, indent=2)
-    return output_path, readiness
-
-
-def build_dataset_quality(pipeline_metadata, hub_health, source_readiness):
-    """Construye tarjeta de puntuación de calidad multidimensional por dataset."""
-    health_by_dataset = {entry["dataset"]: entry for entry in hub_health.get("datasets", [])}
-    readiness_by_dataset = {
-        entry["dataset"]: entry for entry in source_readiness.get("datasets", [])
-    }
-
-    # Ponderaciones
-    WEIGHTS = {
-        "validation": 25,
-        "schema_contract": 20,
-        "source_readiness": 20,
-        "freshness": 15,
-        "coverage": 10,
-        "reuse_policy": 10,
-    }
-
-    def _grade(score):
-        if score >= 90:
-            return "A"
-        if score >= 75:
-            return "B"
-        if score >= 60:
-            return "C"
-        if score >= 40:
-            return "D"
-        return "F"
-
-    def _score_validation(health):
-        return 100 if health.get("validation_status") == "ok" else 0
-
-    def _score_schema_contract(dataset_name):
-        contract_path = os.path.join(
-            ROOT_DIR, "contracts", "datasets", f"{dataset_name}.schema.json"
-        )
-        return 100 if os.path.exists(contract_path) else 0
-
-    def _score_source_readiness(readiness):
-        if readiness.get("live_ready") and readiness.get("live_extractor_status") == "implemented":
-            return 100
-        if readiness.get("fallback_allowed"):
-            return 50
-        return 0
-
-    def _score_freshness(health):
-        status = health.get("freshness_status", "unknown")
-        if status == "fresh":
-            return 100
-        if status == "stale":
-            return 50
-        return 0
-
-    def _score_coverage(health):
-        status = health.get("coverage_status", "unknown")
-        if status == "full":
-            return 100
-        if status == "partial":
-            return 70
-        return 0
-
-    def _score_reuse(health):
-        status = health.get("reuse_status", "unknown")
-        if status == "open-attribution":
-            return 100
-        if status == "public-api-review-terms":
-            return 50
-        return 0
-
-    entries = []
-
-    for dataset_name in sorted(set(list(health_by_dataset) + list(readiness_by_dataset))):
-        health = health_by_dataset.get(dataset_name, {})
-        readiness = readiness_by_dataset.get(dataset_name, {})
-
-        dims = {
-            "validation": _score_validation(health),
-            "schema_contract": _score_schema_contract(dataset_name),
-            "source_readiness": _score_source_readiness(readiness),
-            "freshness": _score_freshness(health),
-            "coverage": _score_coverage(health),
-            "reuse_policy": _score_reuse(health),
-        }
-
-        overall = sum(dims[k] * WEIGHTS[k] for k in WEIGHTS) / 100
-
-        blocking_reasons = []
-        if dims["validation"] < 100:
-            blocking_reasons.append(f"Validación: {health.get('validation_status', 'error')}")
-        if dims["schema_contract"] < 100:
-            blocking_reasons.append("Contrato de esquema ausente")
-        if dims["source_readiness"] < 100:
-            rd = readiness
-            if rd.get("live_extractor_status") == "fallback_only":
-                blocking_reasons.append(
-                    f"Fuente en fallback_only — {rd.get('next_action', 'sin acción definida')}"
-                )
-            elif not rd.get("live_ready"):
-                blocking_reasons.append("Fuente no lista para live")
-            else:
-                blocking_reasons.append("Madurez de fuente incompleta")
-        if dims["freshness"] < 100:
-            blocking_reasons.append(f"Datos {health.get('freshness_status', 'desconocidos')}")
-        if dims["coverage"] < 100:
-            blocking_reasons.append(f"Cobertura {health.get('coverage_status', 'desconocida')}")
-        if dims["reuse_policy"] < 100:
-            blocking_reasons.append(f"Reutilización: {health.get('reuse_status', 'desconocida')}")
-
-        # Bloqueo por upstreams candidatos en datasets derivados
-        maturity = readiness.get("maturity_status", "unknown")
-        if maturity == "candidate" and readiness.get("access_method") == "derived":
-            blocking_reasons.append("Capa derivada — depende de datasets upstream no publicables")
-
-        recommended_action = readiness.get(
-            "next_action",
-            health.get("recommended_action", "—"),
-        )
-        if not blocking_reasons:
-            recommended_action = "Mantener monitoreo operativo y actualizaciones periódicas."
-
-        entries.append(
-            {
-                "dataset": dataset_name,
-                "maturity_status": maturity,
-                "overall_score": round(overall, 1),
-                "grade": _grade(overall),
-                "dimensions": dims,
-                "weights": WEIGHTS,
-                "blocking_reasons": blocking_reasons,
-                "recommended_action": recommended_action,
-            }
-        )
-
-    # Aggregate stats
-    grades = {}
-    for e in entries:
-        g = e["grade"]
-        grades[g] = grades.get(g, 0) + 1
-
-    return {
-        "generated_at_utc": pipeline_metadata.get(
-            "generated_at_utc", datetime.now(UTC).isoformat()
-        ),
-        "dataset_count": len(entries),
-        "grade_distribution": {
-            "A": grades.get("A", 0),
-            "B": grades.get("B", 0),
-            "C": grades.get("C", 0),
-            "D": grades.get("D", 0),
-            "F": grades.get("F", 0),
-        },
-        "average_score": round(sum(e["overall_score"] for e in entries) / len(entries), 1)
-        if entries
-        else 0,
-        "weights": WEIGHTS,
-        "datasets": entries,
-    }
-
-
-def write_dataset_quality_json(quality):
-    output_path = os.path.join(NORMALIZED_DIR, "dataset_quality.json")
-    write_json_atomic(quality, output_path, ensure_ascii=False, indent=2)
-    return output_path, quality
-
-
-def write_hub_bundle_json(pipeline_metadata, hub_health, dataset_catalog, artifact_manifest):
-    artifacts_by_dataset = {}
-    shared_artifacts = []
-
-    for artifact in artifact_manifest.get("artifacts", []):
-        dataset_name = artifact.get("dataset")
-        if dataset_name:
-            artifacts_by_dataset.setdefault(dataset_name, []).append(artifact)
-        else:
-            shared_artifacts.append(artifact)
-
-    shared_artifacts_by_semantic_key = {}
-    for artifact in shared_artifacts:
-        shared_type = artifact.get("shared_type")
-        artifact_format = artifact.get("format")
-        if shared_type and artifact_format:
-            shared_artifacts_by_semantic_key[f"{shared_type}:{artifact_format}"] = artifact
-
-    report_specs = {
-        "status_markdown": ("pipeline_status", "markdown"),
-        "health_json": ("hub_health", "json"),
-        "health_markdown": ("hub_health", "markdown"),
-        "status_json": ("hub_status", "json"),
-        "dataset_status_json": ("dataset_status", "json"),
-        "dataset_changelog_json": ("dataset_changelog", "json"),
-        "bundle_json": ("hub_bundle", "json"),
-        "redistribution_json": ("redistribution_report", "json"),
-        "redistribution_markdown": ("redistribution_report", "markdown"),
-        "provenance_json": ("provenance_report", "json"),
-        "provenance_markdown": ("provenance_report", "markdown"),
-        "drift_json": ("drift_report", "json"),
-        "drift_markdown": ("drift_report", "markdown"),
-        "overview_json": ("overview", "json"),
-        "overview_markdown": ("overview", "markdown"),
-        "catalog_json": ("dataset_catalog", "json"),
-        "catalog_markdown": ("dataset_catalog", "markdown"),
-        "manifest_json": ("artifact_manifest", "json"),
-        "source_readiness_json": ("source_readiness", "json"),
-        "source_readiness_markdown": ("source_readiness", "markdown"),
-        "dataset_quality_json": ("dataset_quality", "json"),
-        "dataset_quality_markdown": ("dataset_quality", "markdown"),
-    }
-    reports = {}
-    for report_name, (shared_type, artifact_format) in report_specs.items():
-        reports[report_name] = shared_artifacts_by_semantic_key.get(
-            f"{shared_type}:{artifact_format}",
-            {
-                "path": None,
-                "shared_type": shared_type,
-                "format": artifact_format,
-            },
-        )
-
-    # Load registry to separate stable_publishable from candidate datasets
-    registry = load_source_registry()
-    registry_by_dataset = {entry["dataset"]: entry for entry in registry}
-
-    stable_publicable_datasets = {
-        entry["dataset"]
-        for entry in registry
-        if entry.get("publication_track") == "stable_publishable"
-        and entry.get("public_bundle_eligible") is True
-    }
-    candidate_datasets = {
-        entry["dataset"] for entry in registry if entry.get("publication_track") == "candidate"
-    }
-
-    full_catalog_count = dataset_catalog.get("dataset_count")
-    public_dataset_count = len(stable_publicable_datasets)
-    candidate_dataset_count = len(candidate_datasets)
-
-    bundle = {
-        "version": pipeline_metadata.get("version", "unknown"),
-        "generated_at_utc": pipeline_metadata.get("generated_at_utc"),
-        "overall_status": hub_health.get("overall_status"),
-        "dataset_count": full_catalog_count,
-        "public_dataset_count": public_dataset_count,
-        "candidate_dataset_count": candidate_dataset_count,
-        "health": {
-            "ok_count": hub_health.get("ok_count"),
-            "warn_count": hub_health.get("warn_count"),
-            "error_count": hub_health.get("error_count"),
-            "live_count": hub_health.get("live_count"),
-            "fallback_count": hub_health.get("fallback_count"),
-            "stale_count": hub_health.get("stale_count"),
-            "publishable_count": hub_health.get("publishable_count"),
-            "review_terms_count": hub_health.get("review_terms_count"),
-            "unknown_reuse_count": hub_health.get("unknown_reuse_count"),
-            "degraded_count": hub_health.get("degraded_count"),
-            "degradation_warning_count": hub_health.get("degradation_warning_count"),
-            "partial_coverage_count": hub_health.get("partial_coverage_count"),
-            "unknown_coverage_count": hub_health.get("unknown_coverage_count"),
-            "drifted_count": hub_health.get("drifted_count"),
-            "warning_count": hub_health.get("warning_count"),
-            "top_issue": hub_health.get("top_issue"),
-            "top_issue_summary": hub_health.get("top_issue_summary"),
-        },
-        "top_issue": hub_health.get("top_issue"),
-        "top_issue_summary": hub_health.get("top_issue_summary"),
-        "datasets": [],
-        "candidate_datasets": [],
-        "reports": reports,
-        "shared_artifacts": shared_artifacts,
-        "packages": artifact_manifest.get("packages", []),
-    }
-
-    health_by_dataset = {entry["dataset"]: entry for entry in hub_health.get("datasets", [])}
-
-    for dataset in dataset_catalog.get("datasets", []):
-        dataset_name = dataset["dataset"]
-        dataset_health = health_by_dataset.get(dataset_name, {})
-
-        # Si este dataset es un alias de otro, hereda los artefactos del canónico
-        # para que ambos aparezcan en el bundle con los mismos archivos.
-        canonical_name = dataset.get("alias_for") or dataset_name
-
-        if dataset_name in stable_publicable_datasets:
-            bundle["datasets"].append(
-                {
-                    "dataset": dataset_name,
-                    "description": dataset.get("description"),
-                    "source_name": dataset.get("source_name"),
-                    "source_url": dataset.get("source_url"),
-                    "source_mode": dataset.get("source_mode"),
-                    "source_detail": dataset.get("source_detail"),
-                    "refreshed_at_utc": dataset.get("refreshed_at_utc"),
-                    "record_count": dataset.get("record_count"),
-                    "indicator_codes": dataset.get("indicator_codes", []),
-                    "indicator_delivery": dataset.get("indicator_delivery", {}),
-                    "join_keys": dataset.get("join_keys", []),
-                    "confidence_tier": dataset.get("confidence_tier"),
-                    "reuse_policy": dataset.get("reuse_policy", {}),
-                    "validation_status": dataset.get("validation_status"),
-                    "freshness": dataset.get("freshness", {}),
-                    "coverage": dataset.get("coverage", {}),
-                    "warning_count": len(dataset.get("warnings", [])),
-                    "severity": dataset_health.get("severity"),
-                    "publishability_status": dataset_health.get("publishability_status"),
-                    "degradation": dataset.get("degradation", {}),
-                    "drift": dataset.get("drift", {}),
-                    "documentation": dataset.get("documentation"),
-                    "outputs": dataset.get("outputs", {}),
-                    "usage_examples": dataset.get("usage_examples", {}),
-                    "artifacts": artifacts_by_dataset.get(canonical_name, []),
-                }
-            )
-
-        if dataset_name in candidate_datasets:
-            reg_entry = registry_by_dataset.get(dataset_name, {})
-            candidate_entry = {
-                "dataset": dataset_name,
-                "maturity_status": reg_entry.get("maturity_status", "unknown"),
-                "publication_track": "candidate",
-                "public_bundle_eligible": False,
-                "source_mode": dataset.get("source_mode"),
-                "source_detail": dataset.get("source_detail"),
-                "freshness": dataset.get("freshness", {}),
-                "next_action": reg_entry.get("next_action", ""),
-            }
-            if reg_entry.get("upstream_datasets"):
-                candidate_entry["upstream_datasets"] = reg_entry["upstream_datasets"]
-            bundle["candidate_datasets"].append(candidate_entry)
-
-    output_path = os.path.join(NORMALIZED_DIR, "hub_bundle.json")
-    write_json_atomic(bundle, output_path, ensure_ascii=False, indent=2)
-    return output_path, bundle
-
-
-def write_publishable_bundle_zip():
-    manifest_path = os.path.join(NORMALIZED_DIR, "artifact_manifest.json")
-    with open(manifest_path, encoding="utf-8") as f:
-        manifest = json.load(f)
-
-    artifacts = manifest.get("artifacts", [])
-    missing = []
-    for artifact in artifacts:
-        relative_path = artifact["path"]
-        absolute_path = os.path.join(DATA_DIR, os.path.relpath(relative_path, "data"))
-        if not os.path.exists(absolute_path):
-            missing.append(relative_path)
-    if missing:
-        summary = ", ".join(missing[:5])
-        suffix = " (y mas)" if len(missing) > 5 else ""
-        raise SystemExit(
-            f"Error: no se puede crear el ZIP; faltan {len(missing)} artefactos: {summary}{suffix}"
-        )
-
-    output_path = os.path.join(NORMALIZED_DIR, PUBLISHABLE_BUNDLE_ZIP_NAME)
-    tmp_path = output_path + ".tmp"
-    with zipfile.ZipFile(tmp_path, "w", compression=zipfile.ZIP_DEFLATED) as archive:
-        for artifact in artifacts:
-            relative_path = artifact["path"]
-            absolute_path = os.path.join(DATA_DIR, os.path.relpath(relative_path, "data"))
-            archive.write(absolute_path, arcname=relative_path)
-
-    with zipfile.ZipFile(tmp_path, "r") as archive:
-        bad_file = archive.testzip()
-        if bad_file is not None:
-            raise SystemExit(f"Error: ZIP corrupto; primer archivo fallido: {bad_file}")
-        if len(archive.namelist()) != len(artifacts):
-            raise SystemExit(
-                f"Error: ZIP incompleto; esperados {len(artifacts)} archivos, "
-                f"encontrados {len(archive.namelist())}"
-            )
-    os.replace(tmp_path, output_path)
-    return output_path
-
-
-def write_publishable_bundle_sha256(zip_path):
-    output_path = os.path.join(NORMALIZED_DIR, PUBLISHABLE_BUNDLE_SHA256_NAME)
-    relative_zip_path = f"data/normalized/{os.path.basename(zip_path)}"
-    sha256 = compute_sha256(zip_path)
-    tmp_path = output_path + ".tmp"
-    with open(tmp_path, "w", encoding="utf-8") as f:
-        f.write(f"{sha256}  {relative_zip_path}\n")
-    os.replace(tmp_path, output_path)
-    return output_path
-
-
-def attach_publishable_package_to_manifest(zip_path, sha256_path, manifest=None):
-    manifest_path = os.path.join(NORMALIZED_DIR, "artifact_manifest.json")
-    if manifest is None:
-        with open(manifest_path, encoding="utf-8") as f:
-            manifest = json.load(f)
-
-    relative_path = f"data/normalized/{os.path.basename(zip_path)}"
-    checksum_path = f"data/normalized/{os.path.basename(sha256_path)}"
-    manifest["packages"] = [
-        {
-            "path": relative_path,
-            "package_type": "zip",
-            "size_bytes": os.path.getsize(zip_path),
-            "sha256": compute_sha256(zip_path),
-            "checksum_algorithm": "sha256",
-            "checksum_path": checksum_path,
-            "verification_command": f"shasum -a 256 -c {checksum_path}",
-        }
-    ]
-
-    write_json_atomic(manifest, manifest_path, ensure_ascii=False, indent=2)
-    return manifest_path, manifest
-
-
-def derive_geography_layers(df_comunas):
-    df_regiones = (
-        df_comunas.select(["codigo_region", "nombre_region"]).unique().sort("codigo_region")
-    )
-    df_provincias = (
-        df_comunas.select(
-            ["codigo_region", "nombre_region", "codigo_provincia", "nombre_provincia"]
-        )
-        .unique()
-        .sort(["codigo_region", "codigo_provincia"])
-    )
-    return df_regiones, df_provincias
-
-
-def build_perfil_territorial_comunal(
-    df_comunas,
-    df_censo,
-    df_censo_hogares,
-    df_salud,
-    df_educacionales,
-    df_electoral,
-    df_finanzas,
-    df_resultados,
-    df_siedu,
-):
-    salud_counts = (
-        df_salud.group_by("codigo_comuna")
-        .agg(pl.len().alias("establecimientos_salud_total"))
-        .with_columns(pl.col("codigo_comuna").cast(pl.String))
-    )
-    educ_counts = (
-        df_educacionales.group_by("codigo_comuna")
-        .agg(pl.len().alias("establecimientos_educacionales_total"))
-        .with_columns(pl.col("codigo_comuna").cast(pl.String))
-    )
-    latest_finanzas = (
-        df_finanzas.sort(["codigo_comuna", "anio"])
-        .group_by("codigo_comuna")
-        .tail(1)
-        .select(
-            "codigo_comuna",
-            pl.col("anio").alias("anio_finanzas"),
-            "ingresos_totales",
-            "gastos_totales",
-            "ingresos_propios_permanentes",
-            "fondo_comun_municipal",
-            "gasto_personal",
-            "gasto_inversion",
-        )
-    )
-    latest_resultados = (
-        df_resultados.sort(["codigo_comuna", "anio"])
-        .group_by("codigo_comuna")
-        .tail(1)
-        .select(
-            "codigo_comuna",
-            pl.col("anio").alias("anio_resultados_educacionales"),
-            "matricula_total",
-            "asistencia_promedio",
-            "tasa_aprobacion",
-            "tasa_reprobacion",
-            "tasa_retiro",
-            "establecimientos_reportados",
-        )
-    )
-    siedu_summary = (
-        df_siedu.group_by("codigo_comuna")
-        .agg(
-            pl.col("codigo_indicador").n_unique().alias("indicadores_siedu_total"),
-            pl.col("valor").mean().alias("valor_promedio_siedu"),
-        )
-        .with_columns(pl.col("codigo_comuna").cast(pl.String))
-    )
-
-    return (
-        df_comunas.join(
-            df_censo.select(
-                "codigo_comuna",
-                "poblacion_censada",
-                pl.col("hombres").alias("poblacion_hombres"),
-                pl.col("mujeres").alias("poblacion_mujeres"),
-                "poblacion_0_14",
-                "poblacion_15_29",
-                "poblacion_30_44",
-                "poblacion_45_64",
-                "poblacion_65_mas",
-            ),
-            on="codigo_comuna",
-            how="left",
-        )
-        .join(
-            df_censo_hogares.select(
-                "codigo_comuna",
-                "viviendas_censadas",
-                "hogares_censados",
-                pl.col("promedio_personas_hogar").alias("promedio_personas_por_hogar"),
-            ),
-            on="codigo_comuna",
-            how="left",
-        )
-        .join(salud_counts, on="codigo_comuna", how="left")
-        .join(educ_counts, on="codigo_comuna", how="left")
-        .join(
-            df_electoral.select(
-                "codigo_comuna", "distrito_electoral", "circunscripcion_senatorial"
-            ),
-            on="codigo_comuna",
-            how="left",
-        )
-        .join(latest_finanzas, on="codigo_comuna", how="left")
-        .join(latest_resultados, on="codigo_comuna", how="left")
-        .join(siedu_summary, on="codigo_comuna", how="left")
-        .with_columns(
-            pl.col("establecimientos_salud_total").fill_null(0).cast(pl.Int64),
-            pl.col("establecimientos_educacionales_total").fill_null(0).cast(pl.Int64),
-            pl.col("indicadores_siedu_total").fill_null(0).cast(pl.Int64),
-        )
-        .sort("codigo_comuna")
-    )
-
-
-def build_duckdb(
-    df_regiones,
-    df_provincias,
-    df_comunas,
-    df_indicadores,
-    df_censo,
-    df_salud,
-    df_educacionales,
-    extra_tables,
-    output_path,
-):
-    print(f"Compilando base de datos DuckDB en: {output_path}")
-    tmp_path = output_path + ".tmp"
-    if os.path.exists(tmp_path):
-        os.remove(tmp_path)
-
-    con = duckdb.connect(tmp_path)
-    try:
-        # Registrar los DataFrames de Polars como vistas temporales en DuckDB
-        con.register("df_regiones_view", df_regiones)
-        con.register("df_provincias_view", df_provincias)
-        con.register("df_comunas_view", df_comunas)
-        con.register("df_indicadores_view", df_indicadores)
-        con.register("df_censo_view", df_censo)
-        con.register("df_salud_view", df_salud)
-        con.register("df_educacionales_view", df_educacionales)
-        for table_name, df_extra in extra_tables.items():
-            con.register(f"df_{table_name}_view", df_extra)
-
-        # Crear tablas físicas en DuckDB
-        con.execute("CREATE TABLE regiones AS SELECT * FROM df_regiones_view")
-        con.execute("CREATE TABLE provincias AS SELECT * FROM df_provincias_view")
-        con.execute("CREATE TABLE comunas AS SELECT * FROM df_comunas_view")
-        con.execute("CREATE VIEW comunas_enriquecidas AS SELECT * FROM comunas")
-        con.execute("CREATE TABLE indicadores AS SELECT * FROM df_indicadores_view")
-        con.execute("CREATE TABLE censo_comunal AS SELECT * FROM df_censo_view")
-        con.execute("CREATE TABLE establecimientos_salud AS SELECT * FROM df_salud_view")
-        con.execute(
-            "CREATE TABLE establecimientos_educacionales AS SELECT * FROM df_educacionales_view"
-        )
-        for table_name in extra_tables:
-            con.execute(f"CREATE TABLE {table_name} AS SELECT * FROM df_{table_name}_view")
-
-        # Agregar índices básicos para mejorar rendimiento en queries
-        con.execute("CREATE UNIQUE INDEX idx_region_code ON regiones (codigo_region)")
-        con.execute("CREATE UNIQUE INDEX idx_provincia_code ON provincias (codigo_provincia)")
-        con.execute("CREATE UNIQUE INDEX idx_comuna_code ON comunas (codigo_comuna)")
-        con.execute("CREATE INDEX idx_indicador_date ON indicadores (fecha, codigo_indicador)")
-        con.execute("CREATE UNIQUE INDEX idx_censo_comuna ON censo_comunal (codigo_comuna)")
-        con.execute("CREATE INDEX idx_salud_comuna ON establecimientos_salud (codigo_comuna)")
-        con.execute("CREATE UNIQUE INDEX idx_educ_rbd ON establecimientos_educacionales (rbd)")
-        con.execute(
-            "CREATE INDEX idx_educ_comuna ON establecimientos_educacionales (codigo_comuna)"
-        )
-        for table_name in extra_tables:
-            if "codigo_comuna" in extra_tables[table_name].columns:
-                con.execute(f"CREATE INDEX idx_{table_name}_comuna ON {table_name} (codigo_comuna)")
-
-        print("Tablas e índices creados con éxito en DuckDB.")
-    finally:
-        con.close()
-
-    os.replace(tmp_path, output_path)
-
-
-def build_sqlite(
-    df_regiones,
-    df_provincias,
-    df_comunas,
-    df_indicadores,
-    df_censo,
-    df_salud,
-    df_educacionales,
-    extra_tables,
-    extra_tables_pd=None,
-    output_path=None,
-):
-    print(f"Compilando base de datos SQLite en: {output_path}")
-    tmp_path = output_path + ".tmp"
-    if os.path.exists(tmp_path):
-        os.remove(tmp_path)
-
-    # Convertimos a Pandas para inserción con to_sql de pandas
-    df_regiones_pd = df_regiones.to_pandas()
-    df_provincias_pd = df_provincias.to_pandas()
-    df_comunas_pd = df_comunas.to_pandas()
-    df_indicadores_pd = df_indicadores.to_pandas()
-    df_censo_pd = df_censo.to_pandas()
-    df_salud_pd = df_salud.to_pandas()
-    df_educacionales_pd = df_educacionales.to_pandas()
-    if extra_tables_pd is None:
-        extra_tables_pd = {name: df.to_pandas() for name, df in extra_tables.items()}
-
-    # SQLite no maneja Date de forma nativa como tipo fecha real (los guarda como string ISO)
-    # Por lo tanto, convertimos las fechas a string ISO antes de guardar
-    df_indicadores_pd["fecha"] = df_indicadores_pd["fecha"].astype(str)
-
-    conn = sqlite3.connect(tmp_path)
-    try:
-        df_regiones_pd.to_sql("regiones", conn, index=False, if_exists="replace")
-        df_provincias_pd.to_sql("provincias", conn, index=False, if_exists="replace")
-        df_comunas_pd.to_sql("comunas", conn, index=False, if_exists="replace")
-        conn.execute("CREATE VIEW comunas_enriquecidas AS SELECT * FROM comunas")
-        df_indicadores_pd.to_sql("indicadores", conn, index=False, if_exists="replace")
-        df_censo_pd.to_sql("censo_comunal", conn, index=False, if_exists="replace")
-        df_salud_pd.to_sql("establecimientos_salud", conn, index=False, if_exists="replace")
-        df_educacionales_pd.to_sql(
-            "establecimientos_educacionales", conn, index=False, if_exists="replace"
-        )
-        # SQLite no es eficiente para tablas masivas.  Omitimos las que
-        # superen el umbral; DuckDB y Parquet cubren ese caso de uso.
-        _SQLITE_MAX_ROWS = 500_000
-        _SQLITE_MAX_VARS = 999
-        for table_name, df_extra in extra_tables_pd.items():
-            num_rows = len(df_extra)
-            if num_rows > _SQLITE_MAX_ROWS:
-                print(
-                    f"  Omite SQLite para {table_name} ({num_rows:,} filas > "
-                    f"{_SQLITE_MAX_ROWS:,}) — usa DuckDB o Parquet.",
-                    flush=True,
-                )
-                continue
-            num_cols = len(df_extra.columns)
-            if num_rows > 10_000 and num_cols > 0:
-                chunksize = _SQLITE_MAX_VARS // num_cols
-                df_extra.to_sql(
-                    table_name,
-                    conn,
-                    index=False,
-                    if_exists="replace",
-                    method="multi",
-                    chunksize=chunksize,
-                )
-            else:
-                df_extra.to_sql(table_name, conn, index=False, if_exists="replace")
-
-        # Crear índices en SQLite
-        cursor = conn.cursor()
-        cursor.execute("CREATE UNIQUE INDEX idx_lite_region ON regiones (codigo_region)")
-        cursor.execute("CREATE UNIQUE INDEX idx_lite_provincia ON provincias (codigo_provincia)")
-        cursor.execute("CREATE UNIQUE INDEX idx_lite_comuna ON comunas (codigo_comuna)")
-        cursor.execute("CREATE INDEX idx_lite_indicador ON indicadores (fecha, codigo_indicador)")
-        cursor.execute("CREATE UNIQUE INDEX idx_lite_censo ON censo_comunal (codigo_comuna)")
-        cursor.execute("CREATE INDEX idx_lite_salud ON establecimientos_salud (codigo_comuna)")
-        cursor.execute(
-            "CREATE UNIQUE INDEX idx_lite_educ_rbd ON establecimientos_educacionales (rbd)"
-        )
-        cursor.execute(
-            "CREATE INDEX idx_lite_educ_comuna ON establecimientos_educacionales (codigo_comuna)"
-        )
-        for table_name, df_extra in extra_tables.items():
-            if "codigo_comuna" in df_extra.columns:
-                cursor.execute(
-                    f"CREATE INDEX idx_lite_{table_name}_comuna ON {table_name} (codigo_comuna)"
-                )
-        conn.commit()
-        print("Tablas e índices creados con éxito en SQLite.")
-    finally:
-        conn.close()
-
-    os.replace(tmp_path, output_path)
-
-
-def build_excel(
-    df_regiones,
-    df_provincias,
-    df_comunas,
-    df_indicadores,
-    df_censo,
-    df_salud,
-    df_educacionales,
-    extra_tables,
-    extra_tables_pd=None,
-    output_path=None,
-):
-    print(f"Generando archivo Excel consolidado para no técnicos en: {output_path}")
-    # Convertir a Pandas para exportar de forma robusta con XlsxWriter
-    df_regiones_pd = df_regiones.to_pandas()
-    df_provincias_pd = df_provincias.to_pandas()
-    df_comunas_pd = df_comunas.to_pandas()
-    df_indicadores_pd = df_indicadores.to_pandas()
-    df_censo_pd = df_censo.to_pandas()
-    df_salud_pd = df_salud.to_pandas()
-    df_educacionales_pd = df_educacionales.to_pandas()
-    if extra_tables_pd is None:
-        extra_tables_pd = {name: df.to_pandas() for name, df in extra_tables.items()}
-
-    # Limpieza visual y formateo para Excel
-    # En Excel, queremos que el Código Comuna siga siendo un string para que no se pierdan los ceros iniciales
-    # XlsxWriter nos permite definir formatos específicos por columna
-    tmp_path = output_path + ".tmp.xlsx"
-    with pd_excel_writer(tmp_path) as writer:
-        df_regiones_pd.to_excel(writer, sheet_name="Regiones", index=False)
-        df_provincias_pd.to_excel(writer, sheet_name="Provincias", index=False)
-        df_comunas_pd.to_excel(writer, sheet_name="Comunas y Regiones", index=False)
-        df_indicadores_pd.to_excel(writer, sheet_name="Indicadores Diarios", index=False)
-        df_censo_pd.to_excel(writer, sheet_name="Censo Comunal", index=False)
-        df_salud_pd.to_excel(writer, sheet_name="Establecimientos Salud", index=False)
-        df_educacionales_pd.to_excel(
-            writer, sheet_name="Establecimientos Educacionales", index=False
-        )
-        # Escribir tablas extra, dividiendo las que excedan el límite de filas de Excel
-        extra_sheet_names = {}  # table_name -> [sheet_names]
-        for table_name, df_extra in extra_tables_pd.items():
-            base_sheet = DATASET_CATALOG_CONFIG[table_name]["outputs"].get(
-                "excel_sheet", table_name[:31]
-            )[:31]
-            num_rows = len(df_extra)
-            if num_rows <= EXCEL_MAX_ROWS:
-                df_extra.to_excel(writer, sheet_name=base_sheet, index=False)
-                extra_sheet_names[table_name] = [base_sheet]
-            else:
-                num_parts = (num_rows + EXCEL_MAX_ROWS - 1) // EXCEL_MAX_ROWS
-                print(
-                    f"  ⚠ {base_sheet}: {num_rows:,} filas exceden el límite de Excel "
-                    f"({EXCEL_MAX_ROWS:,}) — dividiendo en {num_parts} hojas."
-                )
-                part_names = []
-                for i in range(num_parts):
-                    start = i * EXCEL_MAX_ROWS
-                    end = min(start + EXCEL_MAX_ROWS, num_rows)
-                    # "HojaBase" → "HojaBase parte 1", …
-                    suffix = f" parte {i + 1}"
-                    max_base = 31 - len(suffix)
-                    part_name = base_sheet[:max_base] + suffix
-                    df_extra.iloc[start:end].to_excel(writer, sheet_name=part_name, index=False)
-                    part_names.append(part_name)
-                extra_sheet_names[table_name] = part_names
-
-        # Acceder a los objetos workbook y worksheet para aplicar formato estético
-        worksheet_regiones = writer.sheets["Regiones"]
-        worksheet_provincias = writer.sheets["Provincias"]
-        workbook = writer.book
-        worksheet_comunas = writer.sheets["Comunas y Regiones"]
-        worksheet_indicadores = writer.sheets["Indicadores Diarios"]
-        worksheet_censo = writer.sheets["Censo Comunal"]
-        worksheet_salud = writer.sheets["Establecimientos Salud"]
-        worksheet_educacionales = writer.sheets["Establecimientos Educacionales"]
-
-        # Formato de texto para el código comunal para prevenir pérdida de ceros
-        text_format = workbook.add_format({"num_format": "@"})
-        worksheet_regiones.set_column("A:A", 12, text_format)
-        worksheet_provincias.set_column("A:A", 12, text_format)
-        worksheet_provincias.set_column("C:C", 15, text_format)
-        worksheet_comunas.set_column("A:A", 12, text_format)  # Código Comuna
-        worksheet_comunas.set_column("D:D", 15, text_format)  # Código Provincia
-        worksheet_comunas.set_column("F:F", 12, text_format)  # Código Región
-
-        # Ajustar anchos de columnas comunes para que sea estético
-        worksheet_comunas.set_column("B:B", 22)  # Nombre Comuna
-        worksheet_comunas.set_column("G:G", 25)  # Nombre Región
-        worksheet_indicadores.set_column("A:A", 15)  # Fecha
-        worksheet_indicadores.set_column("B:B", 18)  # Código Indicador
-        worksheet_indicadores.set_column("C:C", 15)  # Valor
-        worksheet_censo.set_column("A:A", 12, text_format)
-        worksheet_censo.set_column("C:C", 15, text_format)
-        worksheet_censo.set_column("E:E", 15, text_format)
-        worksheet_salud.set_column("A:A", 20, text_format)
-        worksheet_salud.set_column("F:F", 12, text_format)
-        worksheet_salud.set_column("H:H", 15, text_format)
-        worksheet_educacionales.set_column("A:A", 12, text_format)  # rbd
-        worksheet_educacionales.set_column("B:B", 10, text_format)  # dv_rbd
-        worksheet_educacionales.set_column("D:D", 12, text_format)  # codigo_region
-        worksheet_educacionales.set_column("E:E", 12, text_format)  # codigo_comuna
-        for table_name, sheet_names in extra_sheet_names.items():
-            df_extra = extra_tables[table_name]
-            has_codigo_comuna = "codigo_comuna" in df_extra.columns
-            codigo_comuna_idx = (
-                df_extra.columns.index("codigo_comuna") if has_codigo_comuna else None
-            )
-            for sheet_name in sheet_names:
-                worksheet = writer.sheets[sheet_name]
-                if has_codigo_comuna:
-                    worksheet.set_column(codigo_comuna_idx, codigo_comuna_idx, 12, text_format)
-
-    os.replace(tmp_path, output_path)
-    print("Archivo Excel multi-pestaña generado con éxito.")
-
-
-def pd_excel_writer(path):
-    import pandas as pd
-
-    return pd.ExcelWriter(path, engine="xlsxwriter")
-
-
-def write_parquet_atomic(df, path):
-    tmp_path = path + ".tmp"
-    df.write_parquet(tmp_path)
-    os.replace(tmp_path, path)
-
-
-def build_flat_files(
-    df_regiones,
-    df_provincias,
-    df_comunas,
-    df_indicadores,
-    df_censo,
-    df_salud,
-    df_educacionales,
-    extra_tables=None,
-):
-    if extra_tables is None:
-        extra_tables = {}
-    # Generamos archivos Parquet
-    regiones_parquet = os.path.join(NORMALIZED_DIR, "regiones.parquet")
-    provincias_parquet = os.path.join(NORMALIZED_DIR, "provincias.parquet")
-    comunas_parquet = os.path.join(NORMALIZED_DIR, "comunas.parquet")
-    indicadores_parquet = os.path.join(NORMALIZED_DIR, "indicadores.parquet")
-
-    write_parquet_atomic(df_regiones, regiones_parquet)
-    write_parquet_atomic(df_provincias, provincias_parquet)
-    write_parquet_atomic(df_comunas, comunas_parquet)
-    write_parquet_atomic(df_indicadores, indicadores_parquet)
-    write_parquet_atomic(df_censo, os.path.join(NORMALIZED_DIR, "censo_comunal.parquet"))
-    write_parquet_atomic(df_salud, os.path.join(NORMALIZED_DIR, "establecimientos_salud.parquet"))
-    write_parquet_atomic(
-        df_educacionales, os.path.join(NORMALIZED_DIR, "establecimientos_educacionales.parquet")
-    )
-    for table_name, df_extra in extra_tables.items():
-        write_parquet_atomic(df_extra, os.path.join(NORMALIZED_DIR, f"{table_name}.parquet"))
-    print(f"  Archivos Parquet exportados a: {NORMALIZED_DIR}")
-
-    # Generamos los endpoints JSON simulados
-    regiones_json = os.path.join(NORMALIZED_DIR, "regiones.json")
-    provincias_json = os.path.join(NORMALIZED_DIR, "provincias.json")
-    comunas_json = os.path.join(NORMALIZED_DIR, "comunas.json")
-    indicadores_json = os.path.join(NORMALIZED_DIR, "indicadores_hoy.json")
-
-    # Para JSON estáticos orientados a frontend, exportamos como lista de diccionarios
-    # SQLite/DuckDB maneja fechas como objetos datetime.date, por lo que convertimos a str para serialización JSON
-    df_indicadores_serializable = df_indicadores.with_columns(pl.col("fecha").cast(pl.String))
-
-    write_json_atomic(df_regiones.to_dicts(), regiones_json, ensure_ascii=False, indent=2)
-    write_json_atomic(df_provincias.to_dicts(), provincias_json, ensure_ascii=False, indent=2)
-    write_json_atomic(df_comunas.to_dicts(), comunas_json, ensure_ascii=False, indent=2)
-    write_json_atomic(
-        df_indicadores_serializable.to_dicts(), indicadores_json, ensure_ascii=False, indent=2
-    )
-    write_json_atomic(
-        df_censo.to_dicts(),
-        os.path.join(NORMALIZED_DIR, "censo_comunal.json"),
-        ensure_ascii=False,
-        indent=2,
-    )
-    write_json_atomic(
-        df_salud.to_dicts(),
-        os.path.join(NORMALIZED_DIR, "establecimientos_salud.json"),
-        ensure_ascii=False,
-        indent=2,
-    )
-    write_json_atomic(
-        df_educacionales.to_dicts(),
-        os.path.join(NORMALIZED_DIR, "establecimientos_educacionales.json"),
-        ensure_ascii=False,
-        indent=2,
-    )
-    # JSON para tablas extra: omitir las masivas (> 100k filas).
-    # Parquet y DuckDB son los formatos recomendados para grandes volúmenes.
-    _JSON_MAX_ROWS = 100_000
-    for table_name, df_extra in extra_tables.items():
-        num_rows = df_extra.height
-        if num_rows > _JSON_MAX_ROWS:
-            print(
-                f"  Omite JSON para {table_name} ({num_rows:,} filas > {_JSON_MAX_ROWS:,}) — "
-                f"usa Parquet en su lugar."
-            )
-            continue
-        write_json_atomic(
-            df_extra.to_dicts(),
-            os.path.join(NORMALIZED_DIR, f"{table_name}.json"),
-            ensure_ascii=False,
-            indent=2,
-        )
-
-    print(f"  Endpoints JSON exportados a: {NORMALIZED_DIR}")
-
-
-def main():
-    ensure_directories()
+def _load_inputs():
+    """Carga metadatos de staging y lee todos los DataFrames de origen."""
     previous_pipeline_metadata = None
     previous_metadata_path = os.path.join(NORMALIZED_DIR, "pipeline_metadata.json")
     if os.path.exists(previous_metadata_path):
@@ -2415,6 +352,66 @@ def main():
         df_siedu,
     )
 
+    dfs = {
+        "comunas": df_comunas,
+        "indicadores": df_indicadores,
+        "censo": df_censo,
+        "salud": df_salud,
+        "censo_hogares": df_censo_hogares,
+        "electoral": df_electoral,
+        "educacionales": df_educacionales,
+        "finanzas": df_finanzas,
+        "resultados_educacionales": df_resultados_educacionales,
+        "siedu": df_siedu,
+        "empresas": df_empresas,
+        "regiones": df_regiones,
+        "provincias": df_provincias,
+        "perfil_territorial": df_perfil_territorial,
+    }
+    meta = {
+        "comunas": comunas_metadata,
+        "indicadores": indicadores_metadata,
+        "censo": censo_metadata,
+        "salud": salud_metadata,
+        "censo_hogares": censo_hogares_metadata,
+        "electoral": electoral_metadata,
+        "educacionales": educacionales_metadata,
+        "finanzas": finanzas_metadata,
+        "resultados_educacionales": resultados_educacionales_metadata,
+        "siedu": siedu_metadata,
+        "empresas": empresas_metadata,
+    }
+    return dfs, meta, previous_pipeline_metadata
+
+
+def _compute_validations(dfs, meta):
+    """Ejecuta las validaciones de todos los datasets; aborta si alguna falla."""
+    df_comunas = dfs["comunas"]
+    df_indicadores = dfs["indicadores"]
+    df_censo = dfs["censo"]
+    df_salud = dfs["salud"]
+    df_censo_hogares = dfs["censo_hogares"]
+    df_electoral = dfs["electoral"]
+    df_educacionales = dfs["educacionales"]
+    df_finanzas = dfs["finanzas"]
+    df_resultados_educacionales = dfs["resultados_educacionales"]
+    df_siedu = dfs["siedu"]
+    df_empresas = dfs["empresas"]
+    df_regiones = dfs["regiones"]
+    df_provincias = dfs["provincias"]
+    df_perfil_territorial = dfs["perfil_territorial"]
+    comunas_metadata = meta["comunas"]
+    indicadores_metadata = meta["indicadores"]
+    censo_metadata = meta["censo"]
+    salud_metadata = meta["salud"]
+    censo_hogares_metadata = meta["censo_hogares"]
+    electoral_metadata = meta["electoral"]
+    educacionales_metadata = meta["educacionales"]
+    finanzas_metadata = meta["finanzas"]
+    resultados_educacionales_metadata = meta["resultados_educacionales"]
+    siedu_metadata = meta["siedu"]
+    empresas_metadata = meta["empresas"]
+
     validations = {
         "regiones": validate_regiones(df_regiones),
         "provincias": validate_provincias(df_provincias),
@@ -2476,6 +473,25 @@ def main():
             for error in result["errors"]:
                 messages.append(f" - {result['dataset']}: {error}")
         raise SystemExit("\n".join(messages))
+    return validations
+
+
+def _write_data_artifacts(dfs):
+    """Compila DuckDB/SQLite/Excel/archivos planos y parquet/JSON auxiliares."""
+    df_comunas = dfs["comunas"]
+    df_indicadores = dfs["indicadores"]
+    df_censo = dfs["censo"]
+    df_salud = dfs["salud"]
+    df_censo_hogares = dfs["censo_hogares"]
+    df_electoral = dfs["electoral"]
+    df_educacionales = dfs["educacionales"]
+    df_finanzas = dfs["finanzas"]
+    df_resultados_educacionales = dfs["resultados_educacionales"]
+    df_siedu = dfs["siedu"]
+    df_empresas = dfs["empresas"]
+    df_regiones = dfs["regiones"]
+    df_provincias = dfs["provincias"]
+    df_perfil_territorial = dfs["perfil_territorial"]
 
     extra_tables = {
         "finanzas_municipales": df_finanzas,
@@ -2561,246 +577,9 @@ def main():
         indent=2,
     )
 
-    dataset_metadata = {
-        "regiones": {
-            **comunas_metadata,
-            "dataset": "regiones",
-            "record_count": df_regiones.height,
-            "fields": df_regiones.columns,
-            "reuse_policy": DATASET_CATALOG_CONFIG["regiones"]["reuse_policy"],
-            "freshness": build_freshness(
-                comunas_metadata.get("refreshed_at_utc"),
-                DATASET_CATALOG_CONFIG["regiones"]["freshness_policy"]["max_age_hours"],
-            ),
-        },
-        "provincias": {
-            **comunas_metadata,
-            "dataset": "provincias",
-            "record_count": df_provincias.height,
-            "fields": df_provincias.columns,
-            "reuse_policy": DATASET_CATALOG_CONFIG["provincias"]["reuse_policy"],
-            "freshness": build_freshness(
-                comunas_metadata.get("refreshed_at_utc"),
-                DATASET_CATALOG_CONFIG["provincias"]["freshness_policy"]["max_age_hours"],
-            ),
-        },
-        "comunas": {
-            **comunas_metadata,
-            "dataset": "comunas",
-            "record_count": df_comunas.height,
-            "fields": df_comunas.columns,
-            "reuse_policy": DATASET_CATALOG_CONFIG["comunas"]["reuse_policy"],
-            "freshness": build_freshness(
-                comunas_metadata.get("refreshed_at_utc"),
-                DATASET_CATALOG_CONFIG["comunas"]["freshness_policy"]["max_age_hours"],
-            ),
-        },
-        "comunas_enriquecidas": {
-            **comunas_metadata,
-            "dataset": "comunas_enriquecidas",
-            "record_count": df_comunas.height,
-            "fields": df_comunas.columns,
-            "reuse_policy": DATASET_CATALOG_CONFIG["comunas_enriquecidas"]["reuse_policy"],
-            "freshness": build_freshness(
-                comunas_metadata.get("refreshed_at_utc"),
-                DATASET_CATALOG_CONFIG["comunas_enriquecidas"]["freshness_policy"]["max_age_hours"],
-            ),
-        },
-        "indicadores": {
-            **indicadores_metadata,
-            "dataset": "indicadores",
-            "record_count": df_indicadores.height,
-            "fields": df_indicadores.columns,
-            "indicator_codes": sorted(df_indicadores["codigo_indicador"].unique().to_list()),
-            "indicator_delivery": build_indicator_delivery(
-                {
-                    **indicadores_metadata,
-                    "indicator_codes": sorted(
-                        df_indicadores["codigo_indicador"].unique().to_list()
-                    ),
-                }
-            ),
-            "reuse_policy": DATASET_CATALOG_CONFIG["indicadores"]["reuse_policy"],
-            "freshness": build_freshness(
-                indicadores_metadata.get("refreshed_at_utc"),
-                DATASET_CATALOG_CONFIG["indicadores"]["freshness_policy"]["max_age_hours"],
-            ),
-        },
-        "censo_comunal": {
-            **censo_metadata,
-            "dataset": "censo_comunal",
-            "record_count": df_censo.height,
-            "fields": df_censo.columns,
-            "reuse_policy": DATASET_CATALOG_CONFIG["censo_comunal"]["reuse_policy"],
-            "freshness": build_freshness(
-                censo_metadata.get("refreshed_at_utc"),
-                DATASET_CATALOG_CONFIG["censo_comunal"]["freshness_policy"]["max_age_hours"],
-            ),
-        },
-        "establecimientos_salud": {
-            **salud_metadata,
-            "dataset": "establecimientos_salud",
-            "record_count": df_salud.height,
-            "fields": df_salud.columns,
-            "reuse_policy": DATASET_CATALOG_CONFIG["establecimientos_salud"]["reuse_policy"],
-            "freshness": build_freshness(
-                salud_metadata.get("refreshed_at_utc"),
-                DATASET_CATALOG_CONFIG["establecimientos_salud"]["freshness_policy"][
-                    "max_age_hours"
-                ],
-            ),
-        },
-        "establecimientos_educacionales": {
-            **educacionales_metadata,
-            "dataset": "establecimientos_educacionales",
-            "record_count": df_educacionales.height,
-            "fields": df_educacionales.columns,
-            "reuse_policy": DATASET_CATALOG_CONFIG["establecimientos_educacionales"][
-                "reuse_policy"
-            ],
-            "freshness": build_freshness(
-                educacionales_metadata.get("refreshed_at_utc"),
-                DATASET_CATALOG_CONFIG["establecimientos_educacionales"]["freshness_policy"][
-                    "max_age_hours"
-                ],
-            ),
-        },
-        "censo_hogares_viviendas": {
-            **censo_hogares_metadata,
-            "record_count": df_censo_hogares.height,
-            "fields": df_censo_hogares.columns,
-            "freshness": build_freshness(
-                censo_hogares_metadata.get("refreshed_at_utc"),
-                DATASET_CATALOG_CONFIG["censo_hogares_viviendas"]["freshness_policy"][
-                    "max_age_hours"
-                ],
-            ),
-        },
-        "distritos_electorales": {
-            **electoral_metadata,
-            "record_count": df_electoral.height,
-            "fields": df_electoral.columns,
-            "freshness": build_freshness(
-                electoral_metadata.get("refreshed_at_utc"),
-                DATASET_CATALOG_CONFIG["distritos_electorales"]["freshness_policy"][
-                    "max_age_hours"
-                ],
-            ),
-        },
-        "finanzas_municipales": {
-            **finanzas_metadata,
-            "dataset": "finanzas_municipales",
-            "record_count": df_finanzas.height,
-            "fields": df_finanzas.columns,
-            "reuse_policy": DATASET_CATALOG_CONFIG["finanzas_municipales"]["reuse_policy"],
-            "freshness": build_freshness(
-                finanzas_metadata.get("refreshed_at_utc"),
-                DATASET_CATALOG_CONFIG["finanzas_municipales"]["freshness_policy"]["max_age_hours"],
-            ),
-        },
-        "resultados_educacionales": {
-            **resultados_educacionales_metadata,
-            "dataset": "resultados_educacionales",
-            "record_count": df_resultados_educacionales.height,
-            "fields": df_resultados_educacionales.columns,
-            "reuse_policy": DATASET_CATALOG_CONFIG["resultados_educacionales"]["reuse_policy"],
-            "freshness": build_freshness(
-                resultados_educacionales_metadata.get("refreshed_at_utc"),
-                DATASET_CATALOG_CONFIG["resultados_educacionales"]["freshness_policy"][
-                    "max_age_hours"
-                ],
-            ),
-        },
-        "indicadores_urbanos_siedu": {
-            **siedu_metadata,
-            "dataset": "indicadores_urbanos_siedu",
-            "record_count": df_siedu.height,
-            "fields": df_siedu.columns,
-            "reuse_policy": DATASET_CATALOG_CONFIG["indicadores_urbanos_siedu"]["reuse_policy"],
-            "freshness": build_freshness(
-                siedu_metadata.get("refreshed_at_utc"),
-                DATASET_CATALOG_CONFIG["indicadores_urbanos_siedu"]["freshness_policy"][
-                    "max_age_hours"
-                ],
-            ),
-            "coverage": siedu_metadata.get("coverage", {}),
-        },
-        "perfil_territorial_comunal": {
-            "dataset": "perfil_territorial_comunal",
-            "source_name": "chile-hub",
-            "source_url": "https://github.com/cortega26/chile-hub",
-            "source_mode": "live"
-            if all(
-                metadata.get("source_mode") == "live"
-                for metadata in (
-                    comunas_metadata,
-                    censo_metadata,
-                    censo_hogares_metadata,
-                    salud_metadata,
-                    electoral_metadata,
-                    educacionales_metadata,
-                    finanzas_metadata,
-                    resultados_educacionales_metadata,
-                    siedu_metadata,
-                )
-            )
-            else "fallback",
-            "source_detail": "derived_from_validated_chile_hub_layers",
-            "refreshed_at_utc": datetime.now(UTC).isoformat(),
-            "record_count": df_perfil_territorial.height,
-            "fields": df_perfil_territorial.columns,
-            "notes": [
-                "derived_dataset",
-                "upstreams: comunas,censo_comunal,censo_hogares_viviendas,establecimientos_salud,establecimientos_educacionales,distritos_electorales,finanzas_municipales,resultados_educacionales,indicadores_urbanos_siedu",
-            ],
-            "reuse_policy": DATASET_CATALOG_CONFIG["perfil_territorial_comunal"]["reuse_policy"],
-            "freshness": build_freshness(
-                datetime.now(UTC).isoformat(),
-                DATASET_CATALOG_CONFIG["perfil_territorial_comunal"]["freshness_policy"][
-                    "max_age_hours"
-                ],
-            ),
-        },
-        **(
-            {
-                "empresas": {
-                    "dataset": "empresas",
-                    "source_name": empresas_metadata.get("source_name", ""),
-                    "source_url": empresas_metadata.get("source_url", ""),
-                    "source_mode": empresas_metadata.get("source_mode", "fallback"),
-                    "source_detail": empresas_metadata.get(
-                        "source_detail", "datos_gob_cl_ckan_api"
-                    ),
-                    "refreshed_at_utc": empresas_metadata.get("refreshed_at_utc", ""),
-                    "record_count": df_empresas.height,
-                    "fields": df_empresas.columns,
-                    "notes": empresas_metadata.get("notes", []),
-                    "reuse_policy": DATASET_CATALOG_CONFIG["empresas"]["reuse_policy"],
-                    "freshness": build_freshness(
-                        empresas_metadata.get("refreshed_at_utc", ""),
-                        DATASET_CATALOG_CONFIG["empresas"]["freshness_policy"]["max_age_hours"],
-                    ),
-                }
-            }
-            if df_empresas is not None
-            else {}
-        ),
-    }
-    validations_with_freshness = {
-        dataset_name: {
-            **validation,
-            "warnings": validation.get("warnings", [])
-            + build_freshness_warnings(dataset_name, dataset_metadata[dataset_name]["freshness"]),
-            "freshness_status": dataset_metadata[dataset_name]["freshness"]["status"],
-            "freshness_age_hours": dataset_metadata[dataset_name]["freshness"]["age_hours"],
-        }
-        for dataset_name, validation in validations.items()
-    }
-    dataset_metadata = enrich_dataset_metadata(dataset_metadata, validations_with_freshness)
-    metadata_output, pipeline_metadata = write_pipeline_metadata(
-        dataset_metadata,
-        validations_with_freshness,
-    )
+
+def _generate_reports(pipeline_metadata, previous_pipeline_metadata, metadata_output):
+    """Genera reportes, estados, bundle y manifiesto, e imprime el resumen."""
     hub_health = build_hub_health(pipeline_metadata)
     write_status_markdown_file(pipeline_metadata, health=hub_health)
     hub_health_output = write_hub_health_json(hub_health)
@@ -2860,6 +639,29 @@ def main():
     print(f"ZIP publicable exportado a: {zip_output}")
     print(f"SHA256 publicable exportado a: {sha256_output}")
 
+
+def main():
+    ensure_directories()
+    dfs, meta, previous_pipeline_metadata = _load_inputs()
+    validations = _compute_validations(dfs, meta)
+    _write_data_artifacts(dfs)
+    dataset_metadata = build_dataset_metadata(dfs, meta)
+    validations_with_freshness = {
+        dataset_name: {
+            **validation,
+            "warnings": validation.get("warnings", [])
+            + build_freshness_warnings(dataset_name, dataset_metadata[dataset_name]["freshness"]),
+            "freshness_status": dataset_metadata[dataset_name]["freshness"]["status"],
+            "freshness_age_hours": dataset_metadata[dataset_name]["freshness"]["age_hours"],
+        }
+        for dataset_name, validation in validations.items()
+    }
+    dataset_metadata = enrich_dataset_metadata(dataset_metadata, validations_with_freshness)
+    metadata_output, pipeline_metadata = write_pipeline_metadata(
+        dataset_metadata,
+        validations_with_freshness,
+    )
+    _generate_reports(pipeline_metadata, previous_pipeline_metadata, metadata_output)
     print("\n--- Compilación del Sprint 0 completada con éxito ---")
